### 2026-02-16 #10043 @editorialbot
[PRE REVIEW]: infDB: An Open-Source Data Ecosystem for Urban Energy Infrastructure Modelling and Planning
**Submitting author:** @patrick-buchenberg (<a href="http://orcid.org/0000-0001-7683-8422">Patrick Buchenberg</a>)
**Repository:** https://github.com/tum-ens/InfDB
**Branch with paper.md** (empty if default branch): main
**Version:** 3.0.0
