### 2026-02-23 #10043 @patrick-buchenberg
I renamed the reseach impact statement so that should be fine now and also fixed the pdf generation.
