### 2026-03-25 57cec97
adapted to the tools template
