+++ docs/mkdocs/docs/tools/infdb-basedata-ways/architecture.md
@@ -1,0 +1,6 @@
+# Architecture
+
+The architecture of the `infDB-basedata-ways` tool is based on the InfDB Dev Container template as described in the [Tool Template](../dev-container/index.md). The tool is implemented in Python and utilizes the `pyinfdb` package for interaction with InfDB. The database operations are executed using SQL queries.
+
+!!! note "Remark"
+    The architecture follows the InfDB Dev Container template. For detailed structure and dependencies, see the [Tool Template](../dev-container/index.md) section.
