### 2026-03-25 77ebac1
function explanations are added
