+++ docs/mkdocs/docs/tools/infdb-basedata-ways/architecture.md
@@ -13,34 +13,31 @@ The `infDB-basedata-ways` tool transforms raw ways-related input data into struc
 
 The tool reads its source data from existing input tables:
 
-- `opendata.basemap_verkehrslinie`
-- `basedata.buildings`
+- `opendata.basemap_verkehrslinie`: Stores the raw way geometries and related source attributes used as the basis for road network generation.
+- `basedata.buildings`: Stores the building geometries and attributes for buildings that are assigned and connected to suitable ways during processing.
 
 ### Intermediate Tables
 
 During processing, the tool creates and updates temporary or intermediate tables used in the transformation workflow:
 
-- `ways_tem`
-- `ways_tem_connection`
-- `connection_lines_tem`
+- `ways_tem`: Temporary working table used to process and segment ways for the generation of `basedata.ways_per_junction`.
+- `ways_tem_connection`: Temporary working table used to process ways for the generation of `basedata.ways_per_connection`. In this table, ways are additionally split at building-to-way connection points.
+- `connection_lines_tem`: Temporary table used to store the generated building-to-way connection lines during processing.
 
 ### Output Tables
 
 The tool writes the processed results into global output tables:
 
-- `basedata.ways_per_junction`
-- `basedata.nodes_per_junction`
-- `basedata.ways_per_connection`
-- `basedata.nodes_per_connection`
-- `basedata.connection_lines`
+- `basedata.ways_per_connection`: Stores the final way network designed for the Pylovo tool. In this network, ways are additionally segmented at building-to-way connection points.
+- `basedata.nodes_per_connection`: Stores the nodes of the way network contained in `basedata.ways_per_connection`.
+- `basedata.ways_per_junction`: Stores the final way network designed for the linear heat density tool. In this network, ways are segmented at way junctions but not additionally at building-to-way connection points.
+- `basedata.nodes_per_junction`: Stores the nodes of the way network contained in `basedata.ways_per_junction`.
+- `basedata.connection_lines`: Stores the generated connection lines between buildings and their assigned ways.
 
 ## Functions
 
 The main logic of the tool is implemented through SQL scripts and database functions. These components support assignment updates, segmentation, splitting operations, and connection generation.
 
-#### `update_assigned_way_id`
-
-Updates `buildings.assigned_way_id` based on a mapping table or nearest suitable way.
 
 #### `insert_way_segment`
 
@@ -58,26 +55,39 @@ Generates temporary connection candidates between buildings and assigned ways.
 
 Updates `buildings.assigned_way_id` after way-merging operations.
 
+#### `update_assigned_way_id`
+
+Updates `buildings.assigned_way_id` based on a mapping table or nearest suitable way.
+
 ## Processing Steps
 
 The SQL-based processing is organized into multiple sequential scripts. These scripts cover initialization, filtering, assignment, merging, segmentation, connection generation, and final insertion into output tables.
 
-### SQL Scripts
-
-- `create_temp_table.sql`  
+#### 1. `create_temp_table.sql` 
   Creates the temporary working tables `ways_tem` and `connection_lines_tem` from the input ways data restricted to the selected AGS. It keeps one valid `LINESTRING` per source feature, adds additional columns needed in later processing steps, and creates indexes for efficient spatial and key-based operations.
-- `initialization.sql`  
+#### 2. `initialization.sql`  
   Creates shared database functions and global output tables required for the processing workflow. It also ensures that parallel workers can initialize these resources safely without conflicts.
-- `filter_by_class.sql`  
+#### 3. `filter_by_class.sql`  
[…]
