### 2026-04-15 cb67acd
update developing in container documentation
