### 2026-04-15 fdd63c1
some parts of the basedata-ways documentation are adjusted
