+++ services/infdb-import/src/opendata_bavaria.py
@@ -1,5 +1,7 @@
 import json
 import logging
+import multiprocessing as mp
+import os
 import subprocess
 from pathlib import Path
 from typing import Dict, List
@@ -33,6 +35,236 @@ def _get_gpkg_layers(gpkg: Path) -> list[str]:
         # Expected format: "1: layer_name (Geometry Type)"
         out = subprocess.check_output(["ogrinfo", "-ro", "-q", str(gpkg)], text=True, stderr=subprocess.STDOUT)
         return [line.split(":", 1)[1].split("(")[0].strip() for line in out.splitlines() if ":" in line and "(" in line]
+
+
+def _urls_to_local_paths(
+    urls: list[str],
+    base_dir: str | Path,
+    log,
+    *,
+    expected_suffix: str | None = None,
+) -> list[str]:
+    """Resolve current-run URLs to existing local file paths in base_dir."""
+    base_dir = str(base_dir)
+    local_files = []
+    missing_files = []
+
+    for url in urls:
+        filename = os.path.basename(url)
+        local_path = os.path.join(base_dir, filename)
+
+        if expected_suffix and not filename.lower().endswith(expected_suffix.lower()):
+            log.warning("Skipping unexpected file extension for URL: %s", url)
+            continue
+
+        if os.path.isfile(local_path):
+            local_files.append(local_path)
+        else:
+            missing_files.append(local_path)
+
+    if missing_files:
+        log.warning("Expected %d downloaded files are missing locally.", len(missing_files))
+        for path in missing_files[:20]:
+            log.warning("Missing file: %s", path)
+        if len(missing_files) > 20:
+            log.warning("... and %d more missing files.", len(missing_files) - 20)
+
+    return sorted(set(local_files))
+
+
+def _chunk_list(items: list[str], chunk_size: int) -> list[list[str]]:
+    """Split items into fixed-size chunks."""
+    if chunk_size <= 0:
+        raise ValueError("chunk_size must be > 0")
+    return [items[i:i + chunk_size] for i in range(0, len(items), chunk_size)]
+
+
+def _write_file_list(file_paths: list[str], list_path: Path) -> None:
+    """Write one file path per line for GDAL input_file_list usage."""
+    with open(list_path, "w", encoding="utf-8") as f:
+        for path in file_paths:
+            f.write(f"{path}\n")
+
+
+def _cleanup_paths(paths: list[Path], log) -> None:
+    """Delete generated helper/output files if they exist."""
+    for path in paths:
+        try:
+            path.unlink(missing_ok=True)
+        except Exception:
+            log.warning("Could not remove temporary file: %s", path)
+
+
+def _process_dgm1_batch_and_append(
+    batch_files: list[str],
+    tool_name: str,
+    source_srid: int,
+    target_res: float,
+    target_table: str,
+    mask_path_str: str,
+    work_dir: str,
+    batch_index: int,
+    total_batches: int,
+) -> bool:
+    try:
+        infdb = InfDB(tool_name=tool_name, config_path="../configs/config-infdb-import.yml")
+        log = infdb.get_worker_logger()
+
+        if not batch_files:
+            log.info("DGM1 batch %d/%d: empty batch, skipping.", batch_index, total_batches)
+            return True
+
+        work_dir = Path(work_dir)
+        mask_path = Path(mask_path_str)
+
+        file_list_path = work_dir / f"dgm1_batch_{batch_index:04d}.txt"
+        vrt_path = work_dir / f"dgm1_batch_{batch_index:04d}.vrt"
+        clipped_tif = work_dir / f"dgm1_batch_{batch_index:04d}.tif"
+
[…]
