### 2026-04-21 eba9e7c
Adjust .env handling according to new package options
