+++ tools/pylovo-generation/main.py
@@ -1,72 +1,16 @@
 #!/usr/bin/env python3
 """
-Pylovo Generation Tool - Interactive AGS Selection
+Pylovo Generation Tool
 Generates synthetic low-voltage grid topologies for German municipalities.
+AGS selection is handled externally via the AGS environment variable (set by tools.sh / run_ags.py)
+or via the config file (configs/config-pylovo-generation.yml).
 """
 
 import os
 import subprocess
 import sys
-from typing import List
 
 from pyinfdb import InfDB
-
-
-def query_database(infdb: InfDB, query: str) -> str:
-    """Execute a PostgreSQL query and return the result."""
-    db_params = infdb.get_db_parameters_dict()
-    conn_string = (
-        f"postgresql://{db_params['user']}:{db_params['password']}"
-        f"@{db_params['host']}:{db_params['exposed_port']}/{db_params['db']}"
-    )
-    try:
-        result = subprocess.run(["psql", conn_string, "-t", "-c", query], capture_output=True, text=True, check=True)
-        return result.stdout.strip()
-    except subprocess.CalledProcessError:
-        return ""
-
-
-def get_available_ags(infdb: InfDB) -> List[str]:
-    """Query database for available AGS codes from opendata.scope table."""
-    log = infdb.get_logger()
-    log.info("Fetching available AGS codes from opendata.scope table...")
-    ags_query = 'SELECT "AGS" FROM opendata.scope WHERE "AGS" IS NOT NULL ORDER BY "AGS"'
-    result = query_database(infdb, ags_query)
-    # Parse result and remove leading zeros
-    ags_list = [line.strip().lstrip("0") for line in result.split("\n") if line.strip()]
-    return ags_list
-
-
-def prompt_user_selection(infdb: InfDB, available_ags: List[str]) -> str:
-    """Display available AGS codes and prompt user for selection."""
-    log = infdb.get_logger()
-    # Blue color for the ags prompt
-    BLUE = "\033[94m"
-    RESET = "\033[0m"
-
-    log.info(
-        "Enter AGS codes to generate low voltage grids: Single AGS ➜ 9185149 | "
-        "Multiple AGS ➜ 9185149,9185150 | All AGS ➜ all"
-    )
-    log.info(f"{BLUE}Available municipalities (AGS codes):{RESET}")
-    for ags in available_ags:
-        log.info(f"{BLUE}{ags}{RESET}")
-
-    selection = input(f"{BLUE}➜ Your selection: {RESET}").strip()
-
-    if not selection:
-        log.error("No AGS codes entered")
-        sys.exit(1)
-
-    if selection.lower() == "all":
-        ags_list = ",".join(available_ags)
-        log.info("Selected: All municipalities (%s)", ags_list)
-    else:
-        # Remove leading zeros from user input (in case they copy from database)
-        ags_list = ",".join([ags.strip().lstrip("0") for ags in selection.split(",")])
-        log.info("Selected: %s", ags_list)
-
-    return ags_list
 
 
 def run_pylovo_setup(infdb: InfDB) -> None:
@@ -95,7 +39,7 @@ def run_pylovo_generate(infdb: InfDB, ags_list: str) -> None:
 def main() -> None:
     """Main entry point for pylovo-generation tool."""
     # Load InfDB facade (config + logging)
-    infdb = InfDB(tool_name="pylovo-generation", config_path="configs/conf-pylovo-generation.yml")
+    infdb = InfDB(tool_name="pylovo-generation", config_path="configs/config-pylovo-generation.yml")
     # Logger
     log = infdb.get_logger()
     log.info("Starting %s tool", infdb.get_toolname())
@@ -109,17 +53,21 @@ def main() -> None:
     os.environ["TARGET_SCHEMA"] = output_schema
[…]
