+++ tools/pylovo-generation/compose.yml
@@ -8,7 +8,8 @@ services:
             - ../../.env          # root env (DB creds, etc.)
             - env_pylovo_generation.env
         volumes:
-            - ./configs:/app/configs
+            - ./configs:/app/configs                           # infdb tool configs (config-pylovo-generation.yml etc.)
+            - ./configs/pylovo:/app/configs/pylovo             # pylovo config overrides (PYLOVO_CONFIG_DIR)
             - ./src/pylovo:/app/mnt/pylovo
             - ../../src/infdb_package:/app/mnt/infdb_package
         extra_hosts:
