### 2026-04-21 5d56fbb
Add check for pylovo schema
