### 2026-04-21 #10043 @nkrusch
:wave: hello @GregorBecker, @sampottinger and @FlorianK13

would you be interested in reviewing this submission for JOSS?
