+++ services/infdb-import/src/opendata_bavaria.py
@@ -1,15 +1,12 @@
 import json
 import logging
 import multiprocessing as mp
-import os
 import subprocess
 from pathlib import Path
-from typing import Dict, List
 
-import geopandas as gpd
 from pyinfdb import InfDB
 from sqlalchemy import text
-from shapely import wkt as shapely_wkt
+import shlex
 
 from . import utils
 from .lod2 import _build_urls_for_region
@@ -37,48 +34,6 @@ def _get_gpkg_layers(gpkg: Path) -> list[str]:
         return [line.split(":", 1)[1].split("(")[0].strip() for line in out.splitlines() if ":" in line and "(" in line]
 
 
-def _urls_to_local_paths(
-    urls: list[str],
-    base_dir: str | Path,
-    log,
-    *,
-    expected_suffix: str | None = None,
-) -> list[str]:
-    """Resolve current-run URLs to existing local file paths in base_dir."""
-    base_dir = str(base_dir)
-    local_files = []
-    missing_files = []
-
-    for url in urls:
-        filename = os.path.basename(url)
-        local_path = os.path.join(base_dir, filename)
-
-        if expected_suffix and not filename.lower().endswith(expected_suffix.lower()):
-            log.warning("Skipping unexpected file extension for URL: %s", url)
-            continue
-
-        if os.path.isfile(local_path):
-            local_files.append(local_path)
-        else:
-            missing_files.append(local_path)
-
-    if missing_files:
-        log.warning("Expected %d downloaded files are missing locally.", len(missing_files))
-        for path in missing_files[:20]:
-            log.warning("Missing file: %s", path)
-        if len(missing_files) > 20:
-            log.warning("... and %d more missing files.", len(missing_files) - 20)
-
-    return sorted(set(local_files))
-
-
-def _chunk_list(items: list[str], chunk_size: int) -> list[list[str]]:
-    """Split items into fixed-size chunks."""
-    if chunk_size <= 0:
-        raise ValueError("chunk_size must be > 0")
-    return [items[i:i + chunk_size] for i in range(0, len(items), chunk_size)]
-
-
 def _write_file_list(file_paths: list[str], list_path: Path) -> None:
     """Write one file path per line for GDAL input_file_list usage."""
     with open(list_path, "w", encoding="utf-8") as f:
@@ -95,176 +50,181 @@ def _cleanup_paths(paths: list[Path], log) -> None:
             log.warning("Could not remove temporary file: %s", path)
 
 
-def _process_dgm1_batch_and_append(
-    batch_files: list[str],
+def _build_dgm1_single_ags_raster(
+    ags_code: str,
     tool_name: str,
+    config_path: str,
+    dgm1_region_cfg: dict,
+    dgm1_base_dir: str,
     source_srid: int,
     target_res: float,
-    target_table: str,
-    mask_path_str: str,
-    work_dir: str,
-    batch_index: int,
-    total_batches: int,
-) -> bool:
+) -> str | None:
+    """Build one clipped raster for a single AGS from already-downloaded source tiles."""
+    worker_log = log
+
     try:
-        infdb = InfDB(tool_name=tool_name, config_path="../configs/config-infdb-import.yml")
-        log = infdb.get_worker_logger()
+        infdb = InfDB(tool_name=tool_name, config_path=config_path)
+        worker_log = infdb.get_worker_logger()
 
-        if not batch_files:
-            log.info("DGM1 batch %d/%d: empty batch, skipping.", batch_index, total_batches)
-            return True
+        work_dir = Path(dgm1_base_dir)
 
-        work_dir = Path(work_dir)
-        mask_path = Path(mask_path_str)
+        # Resolve AGS-specific URLs and corresponding local TIFFs
[…]
