### 2026-04-22 e668234
update get_envelop function to handle multiprocessing
