### 2026-04-22 2ed32e6
Fix ags run for pylovo-generation
