+++ tools/pylovo-generation/compose.yml
@@ -10,7 +10,7 @@ services:
         volumes:
             - ./configs:/app/configs                           # infdb tool configs (config-pylovo-generation.yml etc.)
             - ./configs/pylovo:/app/configs/pylovo             # pylovo config overrides (PYLOVO_CONFIG_DIR)
-            - ./src/pylovo:/app/mnt/pylovo
+            - ./src/pylovo:/app/src/pylovo
             - ../../src/infdb_package:/app/mnt/infdb_package
         extra_hosts:
             - "host.docker.internal:host-gateway"
