### 2026-04-23 #10043 @nkrusch
Great, thanks @GregorBecker! The review duration varies by many factors (submission quality, reviewers scheduling, author response time, etc.), thus the projected timeline is advisory. We will proceed based on your availability.
