### 2026-04-23 #10458 @nkrusch
:wave:  @sampottinger and @GregorBecker - Thank you for agreeing to review this submission!

As a reviewer, your first step is to **create a checklist** by entering

as a new comment in this thread. The checklist contain the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied.

The [JOSS review](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention `openjournals/joss-reviews#REVIEW_NUMBER` so that a link is created to this thread (and I can keep an eye on what is happening). Also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

Please feel free to ping me if you have any questions/concerns.
