### 2026-04-23 #10458 @nkrusch
@patrick-buchenberg - please debug the paper rendering locally [using these instructions](https://joss.readthedocs.io/en/latest/paper.html#docker). There seems to be an issue

> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/24840964861?check_suite_focus=true). Problem with affiliations for Martin Stengel, perhaps the affiliations index need quoting?.

Once it is working for you locally, enter as a comment here

This should generate a compiled version of the paper.
