### 2026-04-23 7a1bbe8
Fix affiliation formatting and update section title to 'Research impact statement' in paper.md
