### 2026-04-23 9e8e8dd
Fix affiliation formatting for Laura Kuper in paper.md
