### 2026-04-28 d0678f8
resampling logic is parallelized
