+++ services/infdb-import/src/opendata_bavaria.py
@@ -3,10 +3,19 @@ import logging
 import multiprocessing as mp
 import subprocess
 from pathlib import Path
+from typing import Dict, List
+import sys
 
+<<<<<<< HEAD
 from pyinfdb import InfDB
 from sqlalchemy import text
 import shlex
+=======
+import geopandas as gpd
+from infdb import InfDB
+from sqlalchemy import text
+from shapely import wkt as shapely_wkt
+>>>>>>> f30ea441 (resampling logic is parallelized)
 
 from . import utils
 from .lod2 import _build_urls_for_region
@@ -50,181 +59,252 @@ def _cleanup_paths(paths: list[Path], log) -> None:
             log.warning("Could not remove temporary file: %s", path)
 
 
-def _build_dgm1_single_ags_raster(
-    ags_code: str,
-    tool_name: str,
-    config_path: str,
-    dgm1_region_cfg: dict,
-    dgm1_base_dir: str,
-    source_srid: int,
-    target_res: float,
-) -> str | None:
-    """Build one clipped raster for a single AGS from already-downloaded source tiles."""
-    worker_log = log
-
-    try:
-        infdb = InfDB(tool_name=tool_name, config_path=config_path)
-        worker_log = infdb.get_worker_logger()
-
-        work_dir = Path(dgm1_base_dir)
-
-        # Resolve AGS-specific URLs and corresponding local TIFFs
-        urls = _build_urls_for_single_ags("DGM1 Bavaria", dgm1_region_cfg, ags_code, infdb, worker_log)
-        if not urls:
-            worker_log.info("DGM1 AGS %s: no URLs resolved, skipping.", ags_code)
-            return None
-
-        tile_paths = []
-        for url in urls:
-            tif_path = work_dir / Path(url).name
-            if tif_path.exists() and tif_path.is_file():
-                tile_paths.append(str(tif_path))
-
-        tile_paths = sorted(set(tile_paths))
-
-        if not tile_paths:
-            worker_log.info("DGM1 AGS %s: no local TIFF files found, skipping.", ags_code)
-            return None
-
-        # Resolve AGS-specific exact clip geometry
-        clip_wkt, _, _ = utils.get_clip_geometry(target_crs=source_srid, infdb=infdb, state_prefix=ags_code)
-        if not clip_wkt:
-            worker_log.info("DGM1 AGS %s: no clip geometry found, skipping.", ags_code)
-            return None
-
-        scope_geom = shapely_wkt.loads(clip_wkt)
-
-        file_list_path = work_dir / f"dgm1_{ags_code}.txt"
-        vrt_path = work_dir / f"dgm1_{ags_code}.vrt"
-        mask_path = work_dir / f"mask_dgm1_{ags_code}.gpkg"
-        clipped_tif = work_dir / f"dgm1_{ags_code}.tif"
-
-        try:
-            _write_file_list(tile_paths, file_list_path)
-
-            rc = utils.do_cmd(
-                infdb,
-                [
-                    "gdalbuildvrt",
-                    "-input_file_list",
-                    str(file_list_path),
-                    str(vrt_path),
-                ],
-            )
-            if rc != 0:
-                worker_log.error("DGM1 AGS %s: gdalbuildvrt failed.", ags_code)
-                return None
-
-            gdf = gpd.GeoDataFrame(
-                {"id": [1]},
-                geometry=[scope_geom],
-                crs=f"EPSG:{source_srid}",
-            )
-            gdf.to_file(mask_path, layer="mask", driver="GPKG")
-
-            rc = utils.do_cmd(
-                infdb,
-                [
-                    "gdalwarp",
-                    "-overwrite",
-                    "-of",
-                    "GTiff",
-                    "-co",
[…]
