### 2026-04-29 6d20239
parallel lod2 implementation
