+++ services/infdb-import/src/lod2.py
@@ -4,8 +4,144 @@ import sys
 from infdb import InfDB
 from shapely import wkt as shapely_wkt
 from shapely.geometry import box
+import multiprocessing as mp
 
 from . import utils
+
+
+def _urls_to_local_gml_paths(urls: list[str], gml_path: str, log) -> list[str]:
+    """Resolve current-run URLs to existing local GML file paths."""
+    local_files = []
+    missing_files = []
+
+    for url in urls:
+        filename = os.path.basename(url)
+        local_path = os.path.join(gml_path, filename)
+
+        if os.path.isfile(local_path):
+            local_files.append(local_path)
+        else:
+            missing_files.append(local_path)
+
+    if missing_files:
+        log.warning("LoD2: %d expected files are missing after download.", len(missing_files))
+        for path in missing_files[:20]:
+            log.warning("Missing file: %s", path)
+        if len(missing_files) > 20:
+            log.warning("... and %d more missing files.", len(missing_files) - 20)
+
+    return sorted(set(local_files))
+
+
+def _chunk_list(items: list[str], chunk_size: int) -> list[list[str]]:
+    """Split items into fixed-size chunks."""
+    if chunk_size <= 0:
+        raise ValueError("chunk_size must be > 0")
+    return [items[i:i + chunk_size] for i in range(0, len(items), chunk_size)]
+
+
+def _import_lod2_batch(
+    batch_files: list[str],
+    tool_name: str,
+    batch_index: int,
+    total_batches: int,
+) -> bool:
+    """Import one batch of GML files directly with citydb."""
+    try:
+        infdb = InfDB(tool_name=tool_name, config_path="../configs/config-infdb-import.yml")
+        log = infdb.get_worker_logger()
+
+        if not batch_files:
+            log.info("LoD2 batch %d/%d: empty batch, skipping.", batch_index, total_batches)
+            return True
+
+        source_cfg = [infdb.get_toolname(), "sources", "lod2"]
+        params = infdb.get_db_parameters_dict()
+        import_mode = infdb.get_config_value(source_cfg + ["import-mode"]) or "skip"
+
+        log.info(
+            "LoD2 batch %d/%d: importing %d files.",
+            batch_index,
+            total_batches,
+            len(batch_files),
+        )
+
+        cmd_parts = [
+            "citydb",
+            "import",
+            "citygml",
+            "-H",
+            params["host"],
+            "-d",
+            params["db"],
+            "-u",
+            params["user"],
+            "-p",
+            params["password"],
+            "-P",
+            str(params["exposed_port"]),
+            f"--import-mode={import_mode}",
+            *batch_files,
+        ]
+
+        return_code = utils.do_cmd(infdb, cmd_parts)
+
+        if return_code != 0:
+            log.error(
+                "LoD2 batch %d/%d failed with return code %d.",
+                batch_index,
+                total_batches,
+                return_code,
+            )
+            return False
+
+        log.info("LoD2 batch %d/%d imported successfully.", batch_index, total_batches)
+        return True
+
+    except Exception:
+        if "log" in locals():
+            log.exception("LoD2 batch %d/%d failed unexpectedly.", batch_index, total_batches)
+        return False
+
+
+def _import_lod2_files_in_parallel(
+    infdb: InfDB,
+    gml_files: list[str],
+    batch_size: int = 200,
+    processes: int | None = None,
+) -> bool:
+    """Import current-run GML files in parallel batches."""
+    log = infdb.get_worker_logger()
+
+    if not gml_files:
+        log.warning("LoD2: no GML files to import.")
+        return True
+
[…]
