### 2026-04-29 85e3584
parallel baden-würtenberg lod2 import is implemented
