+++ services/infdb-import/src/lod2.py
@@ -1,5 +1,6 @@
 import os
 import sys
+from zipfile import ZipFile, BadZipFile
 
 from infdb import InfDB
 from shapely import wkt as shapely_wkt
@@ -9,8 +10,17 @@ import multiprocessing as mp
 from . import utils
 
 
-def _urls_to_local_gml_paths(urls: list[str], gml_path: str, log) -> list[str]:
-    """Resolve current-run URLs to existing local GML file paths."""
+def _urls_to_local_citygml_paths(urls: list[str], gml_path: str, log) -> list[str]:
+    """Resolve current-run URLs to existing local CityGML file paths.
+
+    Supports:
+    - direct .gml files
+    - .zip files containing .gml files
+
+    Important:
+    This function only resolves files belonging to the current URL list.
+    It does not import every file in the shared CityGML folder.
+    """
     local_files = []
     missing_files = []
 
@@ -18,10 +28,47 @@ def _urls_to_local_gml_paths(urls: list[str], gml_path: str, log) -> list[str]:
         filename = os.path.basename(url)
         local_path = os.path.join(gml_path, filename)
 
-        if os.path.isfile(local_path):
-            local_files.append(local_path)
-        else:
+        if not os.path.isfile(local_path):
             missing_files.append(local_path)
+            continue
+
+        if filename.lower().endswith(".gml"):
+            local_files.append(local_path)
+            continue
+
+        if filename.lower().endswith(".zip"):
+            try:
+                with ZipFile(local_path, "r") as zf:
+                    gml_members = [
+                        member
+                        for member in zf.namelist()
+                        if member.lower().endswith(".gml")
+                    ]
+
+                    if not gml_members:
+                        log.warning("LoD2 ZIP contains no GML files: %s", local_path)
+                        continue
+
+                    for member in gml_members:
+                        extracted_filename = os.path.basename(member)
+
+                        if not extracted_filename:
+                            continue
+
+                        extracted_path = os.path.join(gml_path, extracted_filename)
+
+                        if not os.path.isfile(extracted_path):
+                            with zf.open(member) as src, open(extracted_path, "wb") as dst:
+                                dst.write(src.read())
+
+                        local_files.append(extracted_path)
+
+            except BadZipFile:
+                log.warning("LoD2 invalid ZIP file: %s", local_path)
+
+            continue
+
+        log.warning("LoD2 unsupported file type: %s", local_path)
 
     if missing_files:
         log.warning("LoD2: %d expected files are missing after download.", len(missing_files))
@@ -37,6 +84,7 @@ def _chunk_list(items: list[str], chunk_size: int) -> list[list[str]]:
     """Split items into fixed-size chunks."""
     if chunk_size <= 0:
         raise ValueError("chunk_size must be > 0")
+
     return [items[i:i + chunk_size] for i in range(0, len(items), chunk_size)]
 
 
@@ -144,15 +192,21 @@ def _import_lod2_files_in_parallel(
     return all(results)
 
 
-def _iter_tile_origins_for_geom(geom, tile_size_m: int):
-    """Yield lower-left tile origin coordinates (in meters) for all tiles
+def _iter_tile_origins_for_geom(
+    geom,
+    tile_size_m: int,
+    offset_x_m: int = 0,
+    offset_y_m: int = 0,
+):
[…]
