+++ services/infdb-import/src/opendata_bavaria.py
@@ -5,6 +5,7 @@ import subprocess
 from pathlib import Path
 from typing import Dict, List
 import sys
+import psycopg2
 
 <<<<<<< HEAD
 from pyinfdb import InfDB
@@ -41,7 +42,6 @@ def _get_gpkg_layers(gpkg: Path) -> list[str]:
         # Expected format: "1: layer_name (Geometry Type)"
         out = subprocess.check_output(["ogrinfo", "-ro", "-q", str(gpkg)], text=True, stderr=subprocess.STDOUT)
         return [line.split(":", 1)[1].split("(")[0].strip() for line in out.splitlines() if ":" in line and "(" in line]
-
 
 def _write_file_list(file_paths: list[str], list_path: Path) -> None:
     """Write one file path per line for GDAL input_file_list usage."""
@@ -101,6 +101,7 @@ def _resample_dgm1_tile(
             "-tr",
             str(target_res),
             str(target_res),
+            "-tap",
             "-srcnodata",
             "-9999",
             "-dstnodata",
@@ -170,6 +171,7 @@ def _resample_dgm1_tile_worker(
             "-tr",
             str(target_res),
             str(target_res),
+            "-tap",
             "-srcnodata",
             "-9999",
             "-dstnodata",
@@ -199,53 +201,179 @@ def _resample_dgm1_tile_worker(
         return False, None, str(err)
 
 
-def _resample_dgm1_batch(
+def _resample_and_import_dgm1_batch(
     batch: list[str],
     resampled_dir_str: str,
+    dgm1_base_dir_str: str,
     source_srid: int,
     target_res: float,
+    schema: str,
+    table_base: str,
+    pgurl: str,
     batch_idx: int,
     total_batches: int,
-) -> tuple[bool, list[str], list[str], int, int]:
+) -> tuple[bool, str | None, str | None, list[str], int, int]:
     resampled_paths: list[str] = []
     errors: list[str] = []
 
     resampled_dir = Path(resampled_dir_str)
+    dgm1_base_dir = Path(dgm1_base_dir_str)
 
-    for source_tif_str in batch:
-        source_tif = Path(source_tif_str)
-        output_tif = resampled_dir / source_tif.name
+    stage_tiled_table = f"{schema}.{table_base}_stage_tiled_{batch_idx:04d}"
+    stage_untiled_table = f"{schema}.{table_base}_stage_untiled_{batch_idx:04d}"
 
-        ok, resampled_path, error = _resample_dgm1_tile_worker(
-            source_tif_str=str(source_tif),
-            output_tif_str=str(output_tif),
-            source_srid=source_srid,
-            target_res=target_res,
+    file_list_path = dgm1_base_dir / f"dgm1_batch_{batch_idx:04d}_files.txt"
+    vrt_path = dgm1_base_dir / f"dgm1_batch_{batch_idx:04d}_{target_res:g}m.vrt"
+
+    try:
+        for source_tif_str in batch:
+            source_tif = Path(source_tif_str)
+            output_tif = resampled_dir / source_tif.name
+
+            ok, resampled_path, error = _resample_dgm1_tile_worker(
+                source_tif_str=str(source_tif),
+                output_tif_str=str(output_tif),
+                source_srid=source_srid,
+                target_res=target_res,
+            )
+
+            if ok and resampled_path:
+                resampled_paths.append(resampled_path)
+            else:
+                errors.append(error or f"Unknown error for {source_tif}")
+
+        if errors:
+            return False, None, None, errors, batch_idx, total_batches
+
+        if not resampled_paths:
+            return False, None, None, [f"No resampled paths for batch {batch_idx}"], batch_idx, total_batches
+
+        _write_file_list(resampled_paths, file_list_path)
+
+        proc = subprocess.run(
[…]
