### 2026-05-05 23e09d3
update paper after first joss review
Co-authored-by: Copilot <copilot@github.com>
