+++ joss/paper.bib
@@ -1,42 +1,328 @@
-% This file was created with Citavi 6.12.0.177
-
-@misc{kwp:2026,
- year = {1/19/2026},
- title = {BMWSB Kommunale W{\"a}rmeplanung - BMWSB},
- url = {https://www.bmwsb.bund.de/DE/stadtentwicklung/klimagerechte-stadtentwicklung/kommunale-waermeplanung/kommunale-waermeplanung_node.html},
- urldate = {1/19/2026}
+@legislation{14aEnWGEinzelnorm,
+  title = {Gesetz über die {{Elektrizitäts-}} und {{Gasversorgung}} ({{Energiewirtschaftsgesetz}} - {{EnWG}})},
+  shorttitle = {Energiewirtschaftsgesetz - {{EnWG}}},
+  journaltitle = {Netzorientierte Steuerung von steuerbaren Verbrauchseinrichtungen und steuerbaren Netzanschlüssen; Festlegungskompetenzen},
+  url = {https://www.gesetze-im-internet.de/enwg_2005/__14a.html},
+  urldate = {2026-04-21},
+  file = {/Users/patrickbuchenberg/Zotero/storage/TADSRRX3/__14a.html}
 }
 
-@article{Open-Data-Strategie:2021,
- year = {1/27/2021},
- abstract = {An innovation strategy for social progress and sustainable growth Federal Cabinet's version, 27 January 2021},
- title = {Data Strategy of the Federal German Government},
- url = {https://www.bundesregierung.de/resource/blob/974430/1950464/39b825e852ca60a299e61da93520bc88/2021-08-12-datenstrategie-en-data.pdf?download=1},
- keywords = {An innovation strategy for social progress and sustainable growth Federal Cabinet's version, 27 January 2021},
- file = {Data Strategy of the Federal German Government - 2021-08:Attachments/Data Strategy of the Federal German Government - 2021-08.pdf:application/pdf}
+@legislation{14dEnWGEinzelnorm,
+  title = {Gesetz über die {{Elektrizitäts-}} und {{Gasversorgung}} ({{Energiewirtschaftsgesetz}} - {{EnWG}})},
+  shorttitle = {Gesetz über die {{Elektrizitäts-}} und {{Gasversorgung}}},
+  journaltitle = {Planung und besondere Bedeutung des Verteilernetzausbaus; Festlegungskompetenz; Verordnungsermächtigung},
+  url = {https://www.gesetze-im-internet.de/enwg_2005/__14d.html},
+  urldate = {2026-04-21},
+  file = {/Users/patrickbuchenberg/Zotero/storage/CNN87NTM/__14d.html}
 }
 
-@misc{14d:2026,
- title = {EnWG §14d},
- year = {1/19/2026},
- url = {https://www.gesetze-im-internet.de/enwg_2005/__14d.html},
- urldate = {1/19/2026}
+@legislation{9EEG2023,
+  title = {Gesetz für den {{Ausbau}} erneuerbarer {{Energien}} ({{Erneuerbare-Energien-Gesetz}} - {{EEG}} 2023)},
+  journaltitle = {Technische Vorgaben},
+  url = {https://www.gesetze-im-internet.de/eeg_2014/__9.html},
+  urldate = {2026-04-21},
+  file = {/Users/patrickbuchenberg/Zotero/storage/BVDMQ5XI/__9.html}
 }
 
-@misc{DB4KWP:2026,
[…]
