+++ joss/paper.md
@@ -1,5 +1,6 @@
 ---
-title: 'InfDB: An Open-Source Data Ecosystem for Urban Energy Infrastructure Modelling and Planning'
+title: 'InfDB: An Open-Source Energy and Infrastructure Data
+Ecosystem for Modeling and Planning'
 tags:
     - Python
     - Docker
