### 2026-05-05 27648b2
fix formatting of title in paper.md
