### 2026-05-05 986a58d
add publication date to DB4KWP project entry in paper.bib
