+++ joss/paper.bib
@@ -186,6 +186,7 @@
   author = {Reiner-Lemoine-Institut},
   url = {https://reiner-lemoine-institut.de/en/project/db4kwp-using-databases-and-knowledge-graphs-to-make-the-results-and-data-of-municipal-heat-planning-comparable-and-usable},
   urldate = {2026-04-23},
+  date = {2026-01-01},
   abstract = {In the DB4KWP project, scientists from the Reiner Lemoine Institute are working with the university to develop a central database that uses tools from the Open Energy Platform to standardize, structure, and make publicly accessible the results and information from municipal heat planning (KWP). The aim is to make data for heating planning comparable, usable in the long term, and coordinatable across regions for use in research, administration, and practice.},
   langid = {british},
   organization = {Reiner-Lemoine-Institut},
