### 2026-05-05 2aee6b9
add publication date to 14dEnWGEinzelnorm entry in paper.bib
