+++ joss/paper.bib
@@ -13,6 +13,7 @@
   journaltitle = {Planung und besondere Bedeutung des Verteilernetzausbaus; Festlegungskompetenz; Verordnungsermächtigung},
   url = {https://www.gesetze-im-internet.de/enwg_2005/__14d.html},
   urldate = {2026-04-21},
+  date = {2026-04-21},
   file = {/Users/patrickbuchenberg/Zotero/storage/CNN87NTM/__14d.html}
 }
 
