### 2026-05-10 a04d1df
Add initial implementation of the infdb-fdw-setup tool, created a separate fdw tool folder, implemented yml config setup for fdw
