+++ tools/infdb-fdw-setup/src/infdb_fdw_setup.py
@@ -1,0 +1,112 @@
+"""
+Foreign Data Wrapper (FDW) setup module for InfDB.
+
+This module sets up PostgreSQL Foreign Data Wrapper to access the central
+opendata schema from a remote InfDB instance without duplicating data.
+"""
+
+from infdb import InfDB
+
+
+def setup_fdw(infdb: InfDB) -> None:
+    """
+    Sets up Foreign Data Wrapper to access the central opendata schema.
+
+    This function:
+    1. Enables the postgres_fdw extension
+    2. Creates a foreign server pointing to the central InfDB instance
+    3. Creates a foreign schema mapping to access the opendata schema
+    4. Grants necessary permissions
+
+    Args:
+        infdb: InfDB handler instance
+    """
+    log = infdb.get_logger()
+
+    # Get FDW configuration
+    fdw_config = infdb.get_config_value([infdb.get_toolname(), "fdw"])
+    central_host = fdw_config["central_db"]["host"]
+    central_port = fdw_config["central_db"]["port"]
+    central_db = fdw_config["central_db"]["db"]
+    central_user = fdw_config["central_db"]["user"]
+    central_password = fdw_config["central_db"]["password"]
+    foreign_schema = fdw_config["foreign_schema"]
+    local_schema = fdw_config["local_schema"]
+
+    log.info(f"Setting up FDW to access '{foreign_schema}' schema from central InfDB instance")
+    log.info(f"Central DB: {central_host}:{central_port}/{central_db}")
+
+    # Connect to local database
+    with infdb.connect() as db:
+        try:
+            # 1. Enable postgres_fdw extension
+            log.info("Enabling postgres_fdw extension...")
+            db.execute_query("CREATE EXTENSION IF NOT EXISTS postgres_fdw;")
+
+            # 2. Create foreign server
+            server_name = "infdb_central_server"
+            log.info(f"Creating foreign server '{server_name}'...")
+            create_server_sql = f"""
+                CREATE SERVER IF NOT EXISTS {server_name}
+                FOREIGN DATA WRAPPER postgres_fdw
+                OPTIONS (
+                    host '{central_host}',
+                    port '{central_port}',
+                    dbname '{central_db}'
+                );
+            """
+            db.execute_query(create_server_sql)
+
+            # 3. Create user mapping
+            log.info("Creating user mapping for foreign server...")
+            create_mapping_sql = f"""
+                CREATE USER MAPPING IF NOT EXISTS FOR CURRENT_USER
+                SERVER {server_name}
+                OPTIONS (
+                    user '{central_user}',
+                    password '{central_password}'
+                );
+            """
+            db.execute_query(create_mapping_sql)
+
+            # 4. Create foreign schema
+            log.info(f"Creating foreign schema '{local_schema}' mapping to remote '{foreign_schema}'...")
+            create_schema_sql = f"""
+                CREATE SCHEMA IF NOT EXISTS {local_schema};
+            """
+            db.execute_query(create_schema_sql)
+
+            # 5. Import foreign schema
+            log.info(f"Importing foreign schema '{foreign_schema}' into local schema '{local_schema}'...")
+            import_schema_sql = f"""
+                IMPORT FOREIGN SCHEMA {foreign_schema}
+                FROM SERVER {server_name}
+                INTO {local_schema};
+            """
+            db.execute_query(import_schema_sql)
+
+            # 6. Grant permissions (optional - adjust as needed)
+            log.info("Granting permissions on foreign schema...")
+            grant_sql = f"""
+                GRANT USAGE ON SCHEMA {local_schema} TO PUBLIC;
[…]
