+++ tools/infdb-fdw-setup/Readme.md
@@ -1,0 +1,232 @@
+# InfDB FDW Setup Tool
+
+> Sets up PostgreSQL Foreign Data Wrapper (FDW) to access the central opendata schema from remote InfDB instances without data duplication.
+
+## Table of Contents
+
+- [Overview](#overview)
+- [Features](#features)
+- [Prerequisites](#prerequisites)
+- [Installation](#installation)
+- [Configuration](#configuration)
+- [Usage](#usage)
+- [Output](#output)
+- [Troubleshooting](#troubleshooting)
+- [License and Citation](#license-and-citation)
+- [Contact](#contact)
+
+## Overview
+
+The InfDB FDW Setup Tool configures PostgreSQL Foreign Data Wrapper to allow new InfDB instances to access the central "opendata" schema remotely. This eliminates the need to duplicate open data across multiple database instances, ensuring data consistency and reducing storage requirements.
+
+**Problem Solved:**
+- Multiple InfDB instances need access to the same open data (buildings, administrative areas, census data, etc.)
+- Traditional approach requires importing the same data into each instance
+- FDW allows transparent access to centralized data without duplication
+
+**Use Cases:**
+- Setting up new InfDB instances that need access to centralized open data
+- Maintaining data consistency across multiple analysis environments
+- Reducing storage costs by centralizing reference data
+
+## Features
+
+- Automated FDW setup with proper error handling and cleanup
+- Configurable connection to central InfDB instance
+- Transparent access to remote opendata schema
+- Permission management for multi-user access
+- Comprehensive logging and troubleshooting
+
+## Prerequisites
+
+Before using this tool, ensure you have:
+
+- infDB instance running (see [infDB documentation](https://infdb.readthedocs.io/))
+- Docker and Docker Compose installed
+- Access to a central InfDB instance with the "opendata" schema populated
+- Network connectivity between local and central InfDB instances
+
+## Installation
+
+### Clone the Repository
+
+This tool is part of the infDB repository:
+
+```bash
+# Already included in infDB under tools/infdb-fdw-setup/
+cd infdb/tools/infdb-fdw-setup
+```
+
+## Configuration
+
+### Configuration File
+
+The tool is configured via `configs/config-infdb-fdw-setup.yml`. Copy the template first:
+
+```bash
+cp configs/config-infdb-fdw-setup.yml.template configs/config-infdb-fdw-setup.yml
+```
+
+### Configuration Options
+
+Edit `configs/config-infdb-fdw-setup.yml` with your settings:
+
+```yaml
+infdb-fdw-setup:
+  config-infdb: "config-infdb.yml"
+  logging:
+    path: "infdb-fdw-setup.log"
+    level: "INFO"  # ERROR, WARNING, INFO, DEBUG
+  hosts:
+    postgres:
+      user: None
+      password: None
+      db: None
+      host: None
+      exposed_port: None
+      epsg: None
+  fdw:
+    central_db:
+      host: "host.docker.internal"  # Central InfDB instance host
+      port: 5432  # Central InfDB instance port
+      db: "postgres"  # Central InfDB database name
+      user: "citydb_user"  # Central InfDB user
+      password: "infdb"  # Central InfDB password
+    foreign_schema: "opendata"  # Schema to import via FDW
+    local_schema: "opendata_fdw"  # Local schema name for FDW mapping
+```
+
+**Configuration Parameters:**
+
+| Parameter | Description | Default | Required |
+|-----------|-------------|---------|----------|
+| `fdw/central_db/host` | Hostname/IP of central InfDB | `host.docker.internal` | Yes |
[…]
