+++ tools/infdb-fdw-setup/.devcontainer/docker-compose.yml
@@ -1,0 +1,28 @@
+version: '3.8'
+services:
+  # Update this to the name of the service you want to work with in your docker-compose.yml file
+  infdb-fdw-setup:
+    # Uncomment if you want to override the service's Dockerfile to one in the .devcontainer 
+    # folder. Note that the path of the Dockerfile and context is relative to the *primary* 
+    # docker-compose.yml file (the first in the devcontainer.json "dockerComposeFile"
+    # array). The sample below assumes your primary file is in the root of your project.
+    #
+    # build:
+    #   context: .
+    #   dockerfile: .devcontainer/Dockerfile
+
+    volumes:
+      # Update this to wherever you want VS Code to mount the folder of your project
+      - .:/workspaces/infdb-fdw-setup:cached
+
+    # Uncomment the next four lines if you will use a ptrace-based debugger like C++, Go, and Rust.
+    # cap_add:
+    #   - SYS_PTRACE
+    # security_opt:
+    #   - seccomp:unconfined
+    env_file:
+      - ../../.env  # repo root .env (dev only)
+
+    # Overrides default command so things don't shut down after the process ends.
+    command: ["/bin/sh", "-c", "uv sync && exec sleep infinity"]
+ 
