+++ tools/infdb-fdw-setup/compose.yml
@@ -1,0 +1,20 @@
+services:
+    infdb-fdw-setup:
+        profiles:
+            - infdb-fdw-setup
+        image: infdb/infdb-fdw-setup:latest
+        build:
+            context: .
+            dockerfile: Dockerfile
+        volumes:
+            # - ./configs:/workspaces/infdb-fdw-setup/configs:ro  # for production
+            - .:/workspaces/infdb-fdw-setup # for development only
+            - ../../src/infdb_package:/workspaces/infdb_package
+            # - /workspaces/infdb-fdw-setup/.venv # Exclude .venv from volume mount
+            - .vscode:/workspaces/.vscode:ro
+        extra_hosts:
+            - "host.docker.internal:host-gateway"
+        env_file:
+            - .env
+        environment:
+            - UV_LINK_MODE=copy
