+++ tools/infdb-fdw-setup/CITATION.cff
@@ -1,0 +1,15 @@
+cff-version: 0.0.0
+message: "If you use this software, please cite it using this metadata."
+authors:
+  - family-names: "Buchenberg"
+    given-names: "Patrick"
+    alias: "@pbuchenberg"
+    affiliation: "Technical University of Munich"
+    orcid: "https://orcid.org/0000-0001-7683-8422"
+title: "infdb-fdw-setup"
+type: software
+license: MIT
+version: 0.2.0
+doi: 
+date-released: 2025-13-05
+url: "https://gitlab.lrz.de/tum-ens/infdb"

[…]
