### 2026-05-12 tum-ens/InfDB#2 @sampottinger
Add top of file comments to shell scripts
### Area of Documentation

Some of your shell scripts are quite essential to the project. This includes things like `infdb.sh`.

### Suggested Changes or Additions

Top of file comment blocks to describe the file at a high level (purpose and, if appropriate, usage) may be helpful in orienting new developers to your project.

### Additional Information

Related to https://github.com/openjournals/joss-reviews/issues/10458
