### 2026-05-13 0a06fcc
JOSS - review Add DEVELOPING guidelines #4 + capitalize Readme.md
