### 2026-05-13 2e5e114
JOSS review - Add code of conduct #3
