+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,18 @@
+# Code of Conduct
+
+As contributors and maintainers of the InfDB project, we pledge to foster an open, welcoming, and harassment-free environment for everyone, regardless of background, experience, or identity.
+
+## Full Policy
+
+To ensure a positive and constructive community, this repository and all associated project spaces are governed by our official code of conduct. We expect all community members—including users, contributors, and maintainers—to adhere to these guidelines when participating in issues, pull requests, discussions, and related events.
+
+Please review the complete and detailed policy here:  
+**[InfDB Code of Conduct](https://tum-ens.github.io/InfDB/development/code-of-conduct/)**
+
+## Reporting Issues
+
+If you experience or witness behavior that violates our code of conduct, please report it immediately by contacting the project team. All reports will be reviewed and investigated promptly and confidentially.
+
+## Contact
+
+For inquiries or to report violations, please refer to the Contact section in [README.md](README.md).
