+++ CONTRIBUTING.md
@@ -2,7 +2,7 @@
 
 Welcome to the InfDB community! We're excited to have you here. Whether you're new to open source or experienced, your contributions help us grow.
 
-Pull requests (PRs) are always welcome, but making a PR is just the start. Please respond to comments and requests for changes to help move the process forward. Skip asking for an issue to be assigned to you on GitHub—send in your PR, explain what you did and ask for a review. It makes collaboration and support much easier. Please follow our Code of Conduct, which applies to all interactions, including issues and PRs.
+Pull requests (PRs) are always welcome, but making a PR is just the start. Please respond to comments and requests for changes to help move the process forward. Skip asking for an issue to be assigned to you on GitHub—send in your PR, explain what you did and ask for a review. It makes collaboration and support much easier. Please follow our [Code of Conduct](https://tum-ens.github.io/InfDB/development/code-of-conduct/), which applies to all interactions, including issues and PRs.
 
 For more, please read https://tum-ens.github.io/InfDB/development/
 
