### 2026-05-14 e4476e3
remove outdated infdb-http + service Readme.md s
