### 2026-05-14 5f27e55
JOSS - review Include example usage into quickstart #7
