+++ README.md
@@ -96,6 +96,15 @@ bash infdb.sh import
 bash infdb.sh stop
 ```
 
+#### Run Tools
+Once InfDB has been successfully started and data has been imported, you can use the integrated tools or develop your own using the provided tool framework to interact with InfDB data. Detailed information on available tools and their usage is provided in the [Tools](https://tum-ens.github.io/InfDB/tools/) section of the documentation.
+
+To run the Linear Heat Density demo, execute:
+```bash
+uv run python3 tools/tools.py -p linear
+```
+Additional information is available in the [Linear Heat Density](https://tum-ens.github.io/InfDB/linear-heat-density/) section of the documentation.
+
 <!-- # Changelog
 
 The changelog is maintained in the [CHANGELOG.md](CHANGELOG.md) file. It lists all changes made to the repository. Follow instructions there to document any updates. -->
