### 2026-05-14 ac5d986
JOSS - review Credit third party open source projects #6
