+++ docs/mkdocs/docs/infdb/services/lizmap.md
@@ -1,0 +1,10 @@
+---
+icon: material/map-search
+---
+# Lizmap
+
+## Configuration
+work in progress
+
+## License
+Lizmap is released under the Mozilla Public License Version 2.0 (MPL-2.0), a weak copyleft Open Source license that allows for both open source and proprietary use with certain conditions.
