### 2026-05-14 d07e9c8
JOSS - review Add some kind of security caution surfaced early in docs #8
