### 2026-05-14 521166c
JOSS review - Potential JOSS paper improvements #1
