+++ joss/paper.bib
@@ -328,3 +328,18 @@ url = {https://github.com/fastapi/fastapi}
   year={2026},
   url={https://opencloud.eu}
 }
+
+@software{githubcopilot,
+  author={{GitHub, Inc.}},
+  title={{GitHub Copilot}},
+  year={2026},
+  url={https://github.com/features/copilot}
+}
+
+@software{googlegemini,
+  author={{Google LLC}},
+  title={{Google Gemini}},
+  year={2026},
+  url={https://gemini.google/about/}
+}
+
