### 2026-05-14 c847471
fix typo JOSS paper
