+++ docs/mkdocs/docs/infdb/services.md
@@ -4,6 +4,6 @@ icon: material/cogs
 # Services
 The InfDB platform provides a suite of essential services designed to facilitate database operation and administration, data handling and visualization, and connectivity. Each preconfigured service can be activated individually to tailor the environment to your specific requirements.
 
-![alt text](services.png)
+![alt text](../assets/img/infdb-overview.png)
 
-For a comprehensive list of integrated services and additional configuration information, see [Services](../infdb/index.md#services) .
+The usage of the services of InfDB is explained in the [Usage -> Services](../../usage/services) section.
