### 2026-05-18 feacab9
MONAI: Solution for #9 - Solution for tum-ens/InfDB#9
