### 2026-05-18 d80fc88
MONAI: Solution for #9 - Bounty: tum-ens/InfDB#9
