+++ docs/TOKEN_REVOCATION_CONFIRMATION.md
@@ -1,0 +1,3 @@
+curl --request DELETE \
+  --header "PRIVATE-TOKEN: <admin_or_user_token>" \
+  "https://gitlab.com/api/v4/personal_access_tokens/12345"
