### 2026-05-18 af305be
Allow a drop of the Schema to create a clean import
