### 2026-05-18 32fdb84
Allow a drop of the Schema to create a clean import
