+++ tools/infdb-fdw-setup/compose.yml
@@ -15,6 +15,7 @@ services:
         extra_hosts:
             - "host.docker.internal:host-gateway"
         env_file:
+            - ../../.env
             - .env
         environment:
             - UV_LINK_MODE=copy
