### 2026-05-18 1c16fb2
add pg_dump instructions for Ubuntu and backup command
