### 2026-05-19 4a3627c
On-demand generation of compose file based on selected profiles
