+++ services/infdb-import/src/utils.py
@@ -979,7 +979,7 @@ def create_building_lod2_table(object_id_prefix: str, infdb: InfDB) -> None:
     Creates the flat building_lod2 table for the specified object_id_prefix
     by filtering the source data based on AGS codes.
 
-    :param object_id_prefix: Object ID prefix (e.g., "DEBY" for Bavaria, "DENW" for North Rhine-Westphalia, "DEBW" for Baden-Württemberg)
+    :param object_id_prefix: Object ID prefix (e.g., "DEBY" for Bavaria, "DENW" for North Rhine-Westphalia, "DEBW" for Baden-Württemberg, "Bensheim" for only Bensheim)
     :type object_id_prefix: str
     :param infdb: instance of InfDB for database access and logging
     :type infdb: InfDB
@@ -991,6 +991,7 @@ def create_building_lod2_table(object_id_prefix: str, infdb: InfDB) -> None:
         create_building_lod2_table("DEBY", infdb)
         create_building_lod2_table("DENW", infdb)
         create_building_lod2_table("DEBW", infdb)
+        create_building_lod2_table("Hesse_Bergstrasse", infdb)  
         return
 
     match object_id_prefix:
@@ -1000,6 +1001,8 @@ def create_building_lod2_table(object_id_prefix: str, infdb: InfDB) -> None:
             ags_id = "05"
         case "DEBW":
             ags_id = "08"
+        case "Hesse_Bergstrasse":
+            ags_id = "06431"
         case _:
             log.error(f"Region {object_id_prefix} not supported for building_lod2.sql")
             sys.exit(1)
