### 2026-05-20 5b142b7
Updated Readme.md for the easier run of db base
