+++ tools/infdb-fdw-setup/Readme.md
@@ -69,7 +69,7 @@ cp configs/config-infdb-fdw-setup.yml.template configs/config-infdb-fdw-setup.ym
 
 ### Configuration Options
 
-Edit `configs/config-infdb-fdw-setup.yml` with your settings:
+Edit `configs/config-infdb-fdw-setup.yml` with your settings. A minimal example:
 
 ```yaml
 infdb-fdw-setup:
@@ -77,32 +77,28 @@ infdb-fdw-setup:
   logging:
     path: "infdb-fdw-setup.log"
     level: "INFO"  # ERROR, WARNING, INFO, DEBUG
-  hosts:
-    postgres:
-      user: None
-      password: None
-      db: None
-      host: None
-      exposed_port: None
-      epsg: None
   fdw:
     central_db:
       host: "host.docker.internal"  # Central InfDB instance host
-      port: 5432  # Central InfDB instance port
-      db: "postgres"  # Central InfDB database name
-      user: "citydb_user"  # Central InfDB user
-      password: "infdb"  # Central InfDB password
-    foreign_schema: "opendata"  # Schema to import via FDW
-    local_schema: "opendata_fdw"  # Local schema name for FDW mapping
+      port: 54328                      # Central InfDB instance port
+      db: "infdb"                     # Central InfDB database name (usually $SERVICES_POSTGRES_DB)
+      user: "citydb_user"             # Central InfDB user
+      password: "infdb"               # Central InfDB password
+    foreign_schema: "opendata"        # Schema to import via FDW
+    local_schema: "opendata_fdw"      # Local schema name for FDW mapping
 ```
+
+Notes:
+- The central database name commonly used in this repository is `infdb` (see `SERVICES_POSTGRES_DB` in the root `.env`). Use the actual name of your central database.
+- Ports and hostnames must be reachable from the machine/container running this tool (e.g. `host.docker.internal` when running Docker on macOS).
 
 **Configuration Parameters:**
 
-| Parameter | Description | Default | Required |
-|-----------|-------------|---------|----------|
+| Parameter | Description | Example / Recommended | Required |
+|-----------|-------------|-----------------------|----------|
 | `fdw/central_db/host` | Hostname/IP of central InfDB | `host.docker.internal` | Yes |
-| `fdw/central_db/port` | Port of central InfDB | `5432` | Yes |
-| `fdw/central_db/db` | Database name on central InfDB | `postgres` | Yes |
+| `fdw/central_db/port` | Port of central InfDB | `54328` | Yes |
+| `fdw/central_db/db` | Database name on central InfDB | `infdb` (or `$SERVICES_POSTGRES_DB`) | Yes |
 | `fdw/central_db/user` | Username for central InfDB | `citydb_user` | Yes |
 | `fdw/central_db/password` | Password for central InfDB | `infdb` | Yes |
 | `fdw/foreign_schema` | Schema name on central DB | `opendata` | Yes |
@@ -135,6 +131,53 @@ docker compose up
 
 - **Standard Mode**: Sets up FDW and imports the complete opendata schema
 - **Custom Mode**: Modify configuration to import specific schemas or tables
+
+Quick Start (recommended)
+-------------------------
+
+From the repository root (recommended):
+
+```bash
+# Run the FDW setup from repo root
+docker compose -f tools/infdb-fdw-setup/compose.yml --profile infdb-fdw-setup up
+```
+
+From inside the tool directory:
+
+```bash
+cd tools/infdb-fdw-setup
+docker compose -f compose.yml --profile infdb-fdw-setup up
+```
+
+Notes:
+- If you run from the tool directory, use the local `compose.yml` path. If you run from the repository root, use the relative path `tools/infdb-fdw-setup/compose.yml`.
[…]
