### 2026-05-21 262247f
rename pyinfdb to PyInfDB in docs
