### 2026-05-21 ddfc2ce
add demo section to usage chapter
