+++ docs/mkdocs/docs/usage/demo.md
@@ -1,0 +1,10 @@
+---
+icon: material-play-circle
+---
+# Demo
+Show demo application on how to use InfDB tools for an use case.
+
+!!! note "Under Development"
+    This documentation section is currently being expanded.
+
+    **Maintainer**: Patrick
