### 2026-05-21 cdfb1b1
fix demo icon bug and add demo to usage contents
