+++ docs/mkdocs/docs/usage/index.md
@@ -11,6 +11,7 @@ To get started with InfDB, you will generally follow this path:
 3.  [**Setup**](setup/index.md): Configure services and parameters.
 4.  [**Deployment**](deployment.md): Launch the InfDB stack.
 5.  [**Services**](services.md): Description of InfDB microservices.
-6.  [**Tools**](tools.md): Software that processes data and is connected to InfDB 
+6.  [**Tools**](tools.md): Software that processes data and is connected to InfDB
+7.  [**Demo**](demo.md): Example applications using InfDB tools.
 
 Select a topic from the navigation menu to begin.
