### 2026-05-21 03f8eaa
correct package name from 'infdb' to 'PyInfDB' in documentation
