+++ docs/mkdocs/docs/infdb/index.md
@@ -24,4 +24,4 @@ The InfDB ecosystem includes a variety of tools designed to handle different asp
 For a comprehensive list of integrated tools and additional information, see **[Tools](../tools/index.md)**.
 
 ## Python Package
-Moreover, there is a python package `infdb` that can be used to interact with the InfDB database and services. It provides functionalities for database connections, logging, configuration management, and utility functions. You can find more information about the package in the **[API -> pyinfdb](../api/pyinfdb/index.md)**.
+Moreover, there is a python package `PyInfDB` that can be used to interact with the InfDB database and services. It provides functionalities for database connections, logging, configuration management, and utility functions. You can find more information about the package in the **[API -> PyInfDB](../api/pyinfdb/index.md)**.
