### 2026-05-21 cff1d78
fix jupyter demo
