### 2026-05-25 2038531
cleaned Readme and added read_only and, status to config params
