### 2026-05-25 9f842a6
check for status and added read_only param
