+++ tools/infdb-fdw-setup/main.py
@@ -20,6 +20,11 @@ def main():
     # Initialize InfDB handler
     infdb = InfDB(tool_name="infdb-fdw-setup", config_path="configs/config-infdb-fdw-setup.yml")
 
+    # only run if config-infdb-fdw-setup.yml status is active
+    if not infdb_fdw_setup.is_active(infdb):
+        print(f"{infdb.get_toolname()} tool is not active. Please check the configuration.")
+        return
+    
     # Start message
     log = infdb.get_logger()
     log.info(f"Starting {infdb.get_toolname()} tool")
