### 2026-05-25 834c24e
added to yml status and read_only param
