+++ tools/infdb-fdw-setup/configs/config-infdb-fdw-setup.yml
@@ -23,11 +23,13 @@ infdb-fdw-setup:
             exposed_port: None
             epsg: None
     fdw:
+        status: active # set to "inactive" to disable FDW setup
         central_db:
             host: "host.docker.internal"  # Central InfDB instance host
-            port: 54328  # Central InfDB instance port (from .env SERVICES_POSTGRES_EXPOSED_PORT)
-            db: "infdb"  # Central InfDB database name (from .env SERVICES_POSTGRES_DB)
+            port: 54328  # Central InfDB instance port
+            db: "infdb"  # Central InfDB database name
             user: "infdb_user"  # Central InfDB user
             password: "infdb"  # Central InfDB password
         foreign_schema: "opendata"  # Schema to import via FDW
         local_schema: "opendata_fdw"  # Local schema name for FDW mapping
+        read_only: true # default is True - set to False to allow write access to foreign tables (not recommended for production)
