### 2026-05-27 4557803
JOSS review - Potential JOSS paper improvements #1
