+++ joss/paper.bib
@@ -343,3 +343,43 @@ url = {https://github.com/fastapi/fastapi}
   url={https://gemini.google/about/}
 }
 
+@software{postgresql,
+  author = {{PostgreSQL Global Development Group}},
+  title  = {PostgreSQL: The World's Most Advanced Open Source Relational Database},
+  url    = {https://www.postgresql.org/},
+  year   = {2026}
+}
+
+@software{postgis,
+  author = {{PostGIS Project Steering Committee}},
+  title  = {PostGIS: Spatial and Geographic Objects for PostgreSQL},
+  url    = {https://postgis.net/},
+  year   = {2026}
+}
+
+@software{timescaledb,
+  author = {{Timescale, Inc.}},
+  title  = {TimescaleDB: Time-Series Data Engineered for PostgreSQL},
+  url    = {https://www.timescale.com/},
+  year   = {2026}
+}
+
+@article{yao20183dcitydb,
+  author  = {Yao, Zhihang and Nagel, Claus and Kunde, Felix and Hudra, Gy{\"o}rgy and Willkomm, Philipp and Donaubauer, Andreas and Adolphi, Thomas and Kolbe, Thomas H.},
+  title   = {3DCityDB - a 3D geodatabase solution for the management, analysis, and visualization of semantic 3D city models based on CityGML},
+  journal = {Open Geospatial Data, Software and Standards},
+  year    = {2018},
+  volume  = {3},
+  number  = {1},
+  doi     = {10.1186/s40965-018-0046-7}
+}
+
+@software{pgRouting,
+  author = {{pgRouting Community}},
+  title  = {pgRouting: Extends the PostGIS / PostgreSQL Geospatial Database to Provide Geospatial Routing Functionality},
+  url    = {https://pgrouting.org/},
+  year   = {2026}
+}
+
+
+
