### 2026-05-27 9cf6187
joss review - Suggest using HTTPS endpoint for quickstart #13
