### 2026-05-27 ec2f2c1
joss review - Credit third party open source projects #6
