+++ CREDITS.md
@@ -1,0 +1,23 @@
+# Credits
+
+This project uses and/or builds on the following open-source software components.
+Refer to the linked files for full license and copyright details.
+
+## Third-Party Components
+
+| Component | License File |
+|---|---|
+| 3DCityDB | https://github.com/3dcitydb/3dcitydb/blob/master/LICENSE |
+| Jupyter Notebook | https://github.com/jupyter/notebook/blob/main/LICENSE |
+| Lizmap | https://github.com/3liz/lizmap-web-client/blob/master/license.txt |
+| OpenCloud | https://github.com/opencloud-eu/opencloud/blob/main/LICENSE |
+| pgAdmin 4 | https://github.com/pgadmin-org/pgadmin4/blob/master/LICENSE |
+| PostGIS | https://github.com/postgis/postgis/blob/master/COPYING |
+| PostgreSQL | https://github.com/postgres/postgres/blob/master/COPYRIGHT |
+| PostgREST | https://github.com/PostgREST/postgrest/blob/main/LICENSE |
+| pygeoapi | https://github.com/geopython/pygeoapi/blob/master/LICENSE.md |
+| pygrouting | https://github.com/pgRouting/pgrouting/blob/develop/LICENSE |
+| QWC2 (QGIS Web Client) | https://github.com/qgis/qwc2/blob/master/LICENSE |
+| TimescaleDB | https://github.com/timescale/timescaledb/blob/main/LICENSE |
+
+
