### 2026-05-27 22e59bb
joss review - Potential JOSS paper improvements #1
