+++ joss/paper.bib
@@ -163,6 +163,7 @@
   title = {{{OpenStreetMap}}},
   author = {{OpenStreetMap contributors}},
   date = {2017},
+  urldate = {2026-05-27},
   url = {https://www.openstreetmap.org}
 }
 
@@ -263,6 +264,7 @@
   author={{netCADservice GmbH}},
   title={flexRM},
   year={2026},
+  urldate = {2026-05-27},
   url={https://www.netcadservice.de/},
   note={Software for municipal heat planning and customer management}
 }
@@ -271,7 +273,9 @@
   author={{Solera GmbH}},
   title={Solarea},
   year={2026},
-  note={Proprietary energy modeling software}
+  note={Proprietary energy modeling software},
+  urldate = {2026-05-27},
+  url = {https://solera.de/}
 }
 
 @software{Ramirez_FastAPI,
@@ -279,6 +283,7 @@ author = {Ramírez, Sebastián},
 license = {MIT},
 title = {{FastAPI}},
 date = {2018-12-05},
+urldate = {2026-05-27},
 url = {https://github.com/fastapi/fastapi}
 }
 
@@ -286,6 +291,7 @@ url = {https://github.com/fastapi/fastapi}
   author={{PostgREST Contributors}},
   title={{PostgREST}},
   year={2026},
+  urldate = {2026-05-27},
   url={https://postgrest.org/},
   note={REST API for any Postgres database}
 }
@@ -294,6 +300,7 @@ url = {https://github.com/fastapi/fastapi}
   title={{pygeoAPI}},
   year={2026},
   organization={Open Source Geospatial Foundation (OSGeo)},
+  urldate = {2026-05-27},
   url={https://pygeoapi.io/}
 }
 
@@ -301,6 +308,7 @@ url = {https://github.com/fastapi/fastapi}
   author={{pgAdmin Development Team}},
   title={{pgAdmin 4}},
   year={2026},
+  urldate = {2026-05-27},
   url={https://www.pgadmin.org/}
 }
 
@@ -319,6 +327,7 @@ url = {https://github.com/fastapi/fastapi}
   author={{QGIS Development Team}},
   organization={Open Source Geospatial Foundation},
   year={2026},
+  urldate = {2026-05-27},
   url={https://qgis.org}
 }
 
@@ -326,6 +335,7 @@ url = {https://github.com/fastapi/fastapi}
   author={{OpenCloud GmbH}},
   title={{OpenCloud}},
   year={2026},
+  urldate = {2026-05-27},
   url={https://opencloud.eu}
 }
 
@@ -333,6 +343,7 @@ url = {https://github.com/fastapi/fastapi}
   author={{GitHub, Inc.}},
   title={{GitHub Copilot}},
   year={2026},
+  urldate = {2026-05-27},
   url={https://github.com/features/copilot}
 }
 
@@ -340,12 +351,14 @@ url = {https://github.com/fastapi/fastapi}
   author={{Google LLC}},
   title={{Google Gemini}},
   year={2026},
+  urldate = {2026-05-27},
   url={https://gemini.google/about/}
 }
 
 @software{postgresql,
   author = {{PostgreSQL Global Development Group}},
   title  = {PostgreSQL: The World's Most Advanced Open Source Relational Database},
+  urldate = {2026-05-27},
   url    = {https://www.postgresql.org/},
   year   = {2026}
 }
@@ -353,6 +366,7 @@ url = {https://github.com/fastapi/fastapi}
 @software{postgis,
   author = {{PostGIS Project Steering Committee}},
   title  = {PostGIS: Spatial and Geographic Objects for PostgreSQL},
+  urldate = {2026-05-27},
   url    = {https://postgis.net/},
[…]
