### 2026-05-27 26b782d
joss review - Compose Profile "core" #14
