### 2026-05-29 fe076bf
Add commented code to sample a subset of buildings for testing
