### 2026-05-29 5a4e211
JOSS review - Clarify where to check output from linear heat density #17
