### 2026-05-29 1b95b98
Remove duplicate image reference and adjust formatting in linear heat density documentation
