+++ docs/mkdocs/docs/linear-heat-density/index.md
@@ -3,6 +3,8 @@
 Within the linear heat density demo, we illustrate how to leverage the InfDB platform to estimate the linear heat density of streets as a key metric for assessing the feasibility and efficiency of district heating networks.
 This use case demonstrates the integration of various data sources and analytical tools within the InfDB ecosystem to derive meaningful metrics for urban energy infrastructure planning.
 
+![alt text](liz-browser.png)
+
 ## Toolchain
 
 ![Linear Heat Density Toolchain](toolchain.png)
@@ -14,8 +16,6 @@ The linear heat density toolchain is implemented through a combination of open-s
 3. The linear heat density is calculated by aggregating the heat demand of buildings along each street segment and dividing it by the length of the street.
 
 ![Toolchain](data-modell.png)
-
-![alt text](liz-browser.png)
 
 ## Run Linear Heat Density
 To run the complete linear heat density toolchain, use the following command:
