### 2026-05-29 88486c8
JOSS review - Issues regarding the documentation #16
