+++ docs/mkdocs/docs/usage/setup/yaml-configuration.md
@@ -6,7 +6,25 @@ icon: material/file-code
 
 The configuration for the opendata import is done via a YAML file (default: `configs/config-infdb-import.yml`) due to its complexity.
 
-For detailed configuration options of the **infdb-import**, refer to the [Services > infdb-import](../../infdb/services/infdb-import.md) documentation.
+For detailed configuration options of the **infdb-import**, refer to the [Services > infdb-import](../services/infdb-import.md) documentation.
+
+When editing the YAML file, keep the following in mind:
+
+- The top-level key `infdb-import` groups all settings for one import instance.
+- Use `status: active` to enable a source or dataset, and `status: not-active` to keep it disabled.
+- Paths may contain placeholders such as `{infdb-import/path/opendata}` or `{infdb-import/name}`; these are resolved by the importer at runtime.
+- `None` values in the database connection section are replaced automatically from the central `config-infdb.yml` file.
+- If you want to connect to a remote database, set the PostgreSQL host, port, user, password, database, and EPSG explicitly.
+- Many sources contain nested sections such as `path`, `datasets`, `layer`, or `types`; only configure the sections that are relevant for the source you want to import.
+- URLs, file names, and layer names must match the structure of the upstream data exactly, otherwise the import will fail.
+
+In practice, a typical configuration flow is:
+
+1. Choose the sources and datasets you need.
+2. Set the instance name and scope to limit the import area.
+3. Define local directories for raw, processed, and temporary files.
+4. Configure database access, either via the central configuration or directly in this file.
+5. Enable the required datasets and adjust source-specific options like resolution, layers, or file templates.
 
 Below is an exemplary excerpt of the `config-infdb-import.yml` file:
 
@@ -23,12 +41,18 @@ Below is an exemplary excerpt of the `config-infdb-import.yml` file:
 #   connection parameters (user, password, db, host, exposed_port, epsg)
 #
 infdb-import:
-    name: "oberh_neub_sonth"  # Name of the infdb-import instance
+    name: "oberh-neub-sonth" #"forchheim"  # Name of the infdb-import instance
     scope:  # AGS (Amtlicher Gemeindeschlüssel)
-        # - "09162000"  # Munich
-        - "09780139"  # Sonthofen
+        # - "09162000"  # Munich (BY)
+        # - "09780139"  # Sonthofen (BY)
+        - "05119000" # Oberhausen (NRW)
+        # - "05117000" # Dümpten (NRW)
+        # - "05158012" # Heiligenhausen (NRW)
+        - "09185149" # Neuburg a. d. Donau (BY)
+        # - "09%" # Bavaria (BY)
+        # - "05%" # North Rhine-Westphalia (NRW)
     multiproccesing: 
-        status: not-active
+        status: active
         max_cores: 2    # max cores since of memory limitations to 2
     config-infdb: "config-infdb.yml" # only filename - change path in ".env" file "CONFIG_INFDB_PATH"
     path:
@@ -46,12 +70,188 @@ infdb-import:
             exposed_port: None
             epsg: None # 3035 (Europe)
         webdav:
-            username: infdb
-            access_token: "letdown subscribe lily catchable landmine sphinx"
+            username: need
+            access_token: "moal taunt tricycle uncrown barista testing"
     sources:
-        # Service configuration examples...
+        package:
+            status: not-active
+            url: http://ds1.need.energy:8123/opendata-neuburg-demo.zip
+            path:
+                base: "{infdb-import/path/base}"
+                processed: "{infdb-import/path/opendata}"
+        need:
+            status: not-active
+            host: "ds1.need.energy"
+            database: need
+            port: 5431
+            user: postgres
[…]
