+++ docs/mkdocs/docs/usage/setup/index.md
@@ -10,4 +10,4 @@ Please select a configuration topic:
 - **[Environment Variables](environment-variables.md)**:<br>Configuring the `.env` file for Docker services (ports, credentials, profiles).
 - **[YAML Configuration](yaml-configuration.md)**: <br>Configuring `config-infdb-import.yml` for data import and processing.
 
-For detailed service-specific configuration, please refer to the [Services](../../infdb/services/index.md) section.
+For detailed service-specific configuration, please refer to the [Services](../services/index.md) section.
