+++ docs/mkdocs/docs/usage/setup/environment-variables.md
@@ -45,7 +45,7 @@ BASE_NAME=infdb-demo # (3)
 # ==============================================================================
 # POSTGRESQL DATABASE (Core Service)
 # ==============================================================================
-# Profile: core
+# Profile: db
 
 # Database name
 SERVICES_POSTGRES_DB=infdb  # (4)
@@ -117,10 +117,10 @@ SERVICES_JUPYTER_EXPOSED_PORT=8888
 SERVICES_JUPYTER_ENABLE_LAB=yes
 
 # Authentication token for Jupyter
-SERVICES_JUPYTER_TOKEN=infdb
+SERVICES_JUPYTER_TOKEN=infdb # (19)
 
 # Path to notebook files
-SERVICES_JUPYTER_PATH_BASE=..src/notebooks/
+SERVICES_JUPYTER_PATH_BASE=../src/notebooks/ # (20)
 
 
 # ==============================================================================
@@ -129,16 +129,16 @@ SERVICES_JUPYTER_PATH_BASE=..src/notebooks/
 # Profile: qwc
 
 # Port for QWC web interface
-SERVICES_QWC_EXPOSED_PORT_GUI=80
+SERVICES_QWC_EXPOSED_PORT_GUI=80 # (21)
 
 # Port for QWC internal database
-SERVICES_QWC_EXPOSED_PORT_DB=5434
+SERVICES_QWC_EXPOSED_PORT_DB=5434 # (22)
 
 # Password for QWC PostgreSQL database
-SERVICES_QWC_POSTGRES_PASSWORD=infdb
+SERVICES_QWC_POSTGRES_PASSWORD=infdb # (23)
 
 # JWT secret key for QWC (change this for production!)
-JWT_SECRET_KEY=change-me-in-production
+JWT_SECRET_KEY=change-me-in-production # (24)
 ```
 
 1.  **Profiles**: By default only the PostgreSQL core is activated. You can activate services by adding the needed profile name to this list.
@@ -155,5 +155,133 @@ JWT_SECRET_KEY=change-me-in-production
 12. **pgAdmin Port**: Port that exposes outside of docker and used to access via browser.
 13. **API Port**: Port that exposes outside of docker and used to communicate with other applications.
 14. **PyGeoAPI Port**: Port that exposes outside of docker and used to communicate with other applications.
-15. **PyGeoAPI Host**: Base url for pygeoAPI local: "localhost", remote: "DOMAIN" or "IP-ADDRESS-OF-REMOTE-HOST".
+15. **PyGeoAPI Host**: Base URL for PyGeoAPI local: `localhost`, remote: `DOMAIN` or `IP-ADDRESS-OF-REMOTE-HOST`.
 16. **PostgREST Port**: Port that exposes outside of docker.
+17. **Jupyter Port**: Port that exposes Jupyter outside of Docker.
+18. **Jupyter Lab**: Enables JupyterLab interface (`yes`/`no`).
+19. **Jupyter Token**: Authentication token for Jupyter access.
+20. **Jupyter Path**: Base path to notebook files.
+21. **QWC GUI Port**: Port that exposes the QWC web interface.
+22. **QWC DB Port**: Port that exposes the QWC internal database.
+23. **QWC DB Password**: Password for QWC PostgreSQL database.
+24. **JWT Secret**: Secret key used by QWC JWT authentication (must be changed for production).
+
+## Detailed configuration guide
+
+Use this section as a practical checklist when editing `.env`.
+
+### 1) Choose service profiles (`COMPOSE_PROFILES`)
+
[…]
