+++ docs/mkdocs/docs/usage/tools.md
@@ -7,5 +7,5 @@ Tools are software that interact with InfDB and process data through standardize
 For a comprehensive list of integrated tools and additional information, see [Tools](../tools/index.md).
 
 <p align="center">
-  <img src="/usage/tools.png" alt="InfDB logo" width="200"/>
+  <img src="tools.png" alt="InfDB logo" width="200"/>
 </p>
