### 2026-06-01 6c325cc
JOSS - review: Test Suite for the infdb Python package #15
