### 2026-06-01 091fef7
JOSS review: Issues regarding the documentation #16
