+++ docs/mkdocs/docs/linear-heat-density/tools/tools.md
@@ -1,14 +1,0 @@
-# Tools
-
-This section provides an overview of the used open source tools for the linear heat density calculation and processing.
-
-## infdb-basedata
-
-
-## ro-heat
-The `ro-heat` tool is used to estimate the building heat demand on a building level. It uses statistical data and building characteristics to calculate the heat demand for each building in the dataset.
-
-## process-streets
-
-
-## linear heat density
