+++ docs/mkdocs/docs/usage/index.md
@@ -10,7 +10,7 @@ To get started with InfDB, you will generally follow this path:
 2.  [**Get Software**](get-software.md): Clone the repository.
 3.  [**Setup**](setup/index.md): Configure services and parameters.
 4.  [**Deployment**](deployment.md): Launch the InfDB stack.
-5.  [**Services**](services.md): Description of InfDB microservices.
+5.  [**Services**](services/index.md): Description of InfDB microservices.
 6.  [**Tools**](tools.md): Software that processes data and is connected to InfDB.
 7.  [**Demo**](demo.md): Example applications using InfDB tools.
 
