+++ .gitlab-ci.yml
@@ -25,7 +25,7 @@ cache:
 # Stages
 stages:
   - setup
-#  - test
+  - test
   - lint
   - docs
 
@@ -40,24 +40,22 @@ setup:
     expire_in: 1 hour
 
 ## Test job
-#test:
-#  stage: test
-#  <<: *install_uv
-#  dependencies:
-#    - setup
-#  script:
-#    - source .venv/bin/activate
-#    - uv pip install pytest pytest-cov
-#    - pytest --cov=src --cov-report=xml --cov-report=html
-#  coverage: '/(?i)total.*? (100(?:\.0+)?\%|[1-9]?\d(?:\.\d+)?\%)$/'
-#  artifacts:
-#    reports:
-#      coverage_report:
-#        coverage_format: cobertura
-#        path: coverage.xml
-#    paths:
-#      - htmlcov/
-#    expire_in: 1 week
+test:
+  stage: test
+  dependencies:
+    - setup
+  script:
+    - source .venv/bin/activate
+    - pytest --cov=src --cov-report=xml --cov-report=html
+  coverage: '/(?i)total.*? (100(?:\.0+)?\%|[1-9]?\d(?:\.\d+)?\%)$/'
+  artifacts:
+    reports:
+      coverage_report:
+        coverage_format: cobertura
+        path: coverage.xml
+    paths:
+      - htmlcov/
+    expire_in: 1 week
 
 pre-commit-hook:
   stage: lint
