### 2026-06-07 34a10b8
JOSS review: Test Suite for the infdb Python package #15 - add uv
