### 2026-06-07 8126a5c
JOSS review: Test Suite for the infdb Python package #15 - add pytest-cov
