+++ pyproject.toml
@@ -44,6 +44,7 @@ docs = [
 ]
 test = [
   "pytest>=8.0",
+  "pytest-cov>=6.0",
   {include-group = "precommit-hook"},
   {include-group = "lint"},
   {include-group = "security-lint"},
