### 2026-06-07 #10458 @nkrusch
@patrick-buchenberg and @GregorBecker,
On my side, the automated tests criteria is still open. Per JOSS policy, the review checklist must be marked checked entirely for the review to count as complete.

Once there are two complete reviews, I will move the submission to the post-review phase.
