### 2026-06-08 06f71ab
change infdb to `InfDB`
