+++ joss/paper.md
@@ -72,7 +72,7 @@ bibliography: paper.bib
 
 It bundles a PostgreSQL [@stonebraker1986design] database extended for geospatial, time-series, and graph data as well as standardized REST [@fielding2000architectural] and OpenAPI standards by the Open Geospatial Consortium (OGC) [@ogc_wfs]. Finally, it offers a configurable import service transforming raw public data into structured, version-controlled schemas.
 
-By decoupling data ingestion, storage, and access from downstream analysis, modeling and planning tools, InfDB reduces data pre-processing effort and enables reproducible, transferable energy system modelling workflows across regions and projects. In this way, it can be used for many different applications in energy system modelling, such as district heating planning, electrical distribution network analysis, or urban energy demand estimation.
+By decoupling data ingestion, storage, and access from downstream analysis, modeling and planning tools, `InfDB` reduces data pre-processing effort and enables reproducible, transferable energy system modelling workflows across regions and projects. In this way, it can be used for many different applications in energy system modelling, such as district heating planning, electrical distribution network analysis, or urban energy demand estimation.
   
 # Statement of need
 The transition to a climate-neutral energy system is a central pillar of energy policy, exemplified by Germany's aim for climate neutrality by 2045. New legislative frameworks, such as the requirement for municipal heat planning [@GesetzFurWarmeplanung] and the requirement for grid expansion plans based on regional transition pathway scenarios defined in the German Energy Industry Act [@14dEnWGEinzelnorm], demand that municipalities and Distribution System Operators (DSOs) process vast amounts of energy and infrastructure data.
@@ -100,16 +100,16 @@ Energy and infrastructure data management is an active field with several existi
 The research relevance of `InfDB` lies in its role as a reusable data infrastructure that supports transparent and reproducible energy system analysis workflows. By separating data management from analysis logic, `InfDB` contributes to several recurring methodological requirements in energy research:
 
 * **Reproducibility:** Containerized deployment and configuration-based data ingestion allow complete data pipelines to be rerun and inspected. This enables studies to be reproduced or updated when new data becomes available, without reimplementing preprocessing steps.
-* **Transferability:** By enforcing structured schemas and standardized access interfaces, InfDB allows data processing and analysis workflows to be reused across regions and projects with minimal adaptation, reducing region-specific reimplementation effort.
-* **Separation of Concerns:** InfDB decouples data acquisition, storage, and access from modeling and analysis code. This allows researchers to develop and test analytical methods independently of changes in input data formats or sources.
+* **Transferability:** By enforcing structured schemas and standardized access interfaces, `InfDB` allows data processing and analysis workflows to be reused across regions and projects with minimal adaptation, reducing region-specific reimplementation effort.
+* **Separation of Concerns:** `InfDB` decouples data acquisition, storage, and access from modeling and analysis code. This allows researchers to develop and test analytical methods independently of changes in input data formats or sources.
 * **Workflow Stability:** Stable database schemas and interfaces provide a consistent foundation for iterative research, supporting comparative studies and sensitivity analyses without requiring repeated adjustments to data handling logic.
-* **Methodological Neutrality:** InfDB does not prescribe modeling approaches, optimization methods, or policy assumptions. Its role is limited to providing structured and accessible data, allowing a wide range of analytical methods to be applied without constraint.
+* **Methodological Neutrality:** `InfDB` does not prescribe modeling approaches, optimization methods, or policy assumptions. Its role is limited to providing structured and accessible data, allowing a wide range of analytical methods to be applied without constraint.
 
 # Software Design
[…]
