### 2026-06-08 9ecfec4
JOSS final paper review fixes
