+++ joss/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'InfDB: An Open-Source Energy and Infrastructure Data Ecosystem for Modeling and Planning'
+title: 'InfDB: An Open Source Energy and Infrastructure Data Ecosystem for Modeling and Planning'
 tags:
     - Python
     - Docker
@@ -68,7 +68,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-`InfDB - Energy and Infrastructure Database` is an open-source, containerized data infrastructure for managing and providing access to heterogeneous energy and infrastructure datasets used in urban and regional energy system studies. 
+`InfDB - Energy and Infrastructure Database` is an open source, containerized data infrastructure for managing and providing access to heterogeneous energy and infrastructure datasets used in urban and regional energy system studies. 
 
 It bundles a PostgreSQL [@stonebraker1986design] database extended for geospatial, time-series, and graph data as well as standardized REST [@fielding2000architectural] and OpenAPI standards by the Open Geospatial Consortium (OGC) [@ogc_wfs]. Finally, it offers a configurable import service transforming raw public data into structured, version-controlled schemas.
 
@@ -92,9 +92,9 @@ Energy and infrastructure data management is an active field with several existi
 
 * **Commercial Platforms:** Tools like **nPro** [@wirtz2023npro], **Solarea** [@solarea_software], and **flexRM** [@flexrm_software] offer robust analytics and user interfaces but are proprietary, limiting transparency and community extension.
 * **Open Source Modeling Frameworks:** Tools like **City Energy Analyst (CEA)** [@jimenofonsecaArchitecturebuildingsystemsCityEnergyAnalystCityEnergyAnalyst2025], **EUReCA** [@pratavieraEUReCAOpensourceUrban2021] and **OpenPlan** [@OpenPlan] focus on modeling and optimization but typically assume preprocessed, structured input data provided externally.
-* **Data Initiatives:** The **NEED project** [@duchonPlatformEcosystemProviding2024] provides a decentralized data hub for synthetic energy data, and **DB4KWP** [@reiner-lemoine-institutDB4KWPUsingDatabases] focuses on ontologies (OEO/OEKG) and naming conventions.
+* **Data Initiatives:** The **NEED project** [@duchonPlatformEcosystemProviding2024] provides a decentralized data hub for synthetic energy data, and **DB4KWP** [@reiner-lemoine-institutDB4KWPUsingDatabases] focuses on ontologies Open Energy Ontology (OEO) and Knowledge Graph (OEKG) naming conventions.
 
-`InfDB` fills the gap between available open data sources and existing energy modelling and planning solutions. While simulation tools like CEA focus on modeling, `InfDB` provides the foundational data infrastructure that these tools require. Moreover, unlike static data repositories, `InfDB` offers a dynamic, service-oriented platform that enables users to deploy local instances, continuously integrate fresh datasets, and seamlessly connect with both commercial and open-source downstream tools. Looking forward, `InfDB` can complement ontology initiatives like DB4KWP by providing a technical implementation layer that transforms conceptual data standards into practical, operational systems.
+`InfDB` fills the gap between available open data sources and existing energy modelling and planning solutions. While simulation tools like CEA focus on modeling, `InfDB` provides the foundational data infrastructure that these tools require. Moreover, unlike static data repositories, `InfDB` offers a dynamic, service-oriented platform that enables users to deploy local instances, continuously integrate fresh datasets, and seamlessly connect with both commercial and open source downstream tools. Looking forward, `InfDB` can complement ontology initiatives like DB4KWP by providing a technical implementation layer that transforms conceptual data standards into practical, operational systems.
 
 # Research impact statement
[…]
