### 2026-06-09 842eb0a
update CITATION.cff to JOSS
