+++ services/infdb-import/src/kwp_nrw_oberhausen.py
@@ -2,7 +2,7 @@ import os
 import sys
 from typing import List, Sequence
 
-from infdb import InfDB
+from pyinfdb import InfDB
 
 from . import utils
 
