### 2026-06-09 b1b4fdc
fix mkdocs pyinfdb change error
