+++ compose.yml
@@ -1,14 +1,0 @@
-name: ${BASE_NAME}
-
-include:
-  - path: services/infdb-db/compose.yml
-  - path: services/infdb-lizmap/docker-compose.yml
-  - path: services/pgadmin/compose.yml
-  - path: services/infdb-api/fastapi/compose.yml
-  - path: services/infdb-api/pygeoapi/compose.yml
-  - path: services/infdb-api/postgrest/compose.yml
-
-networks:
-  infdb_network:
-    name: infdb-${BASE_NAME}_network
-    driver: bridge
