### 2026-06-10 27d5406
Update to pyinfdb in tools pyproject.toml
