+++ tools/ro-heat/pyproject.toml
@@ -8,9 +8,9 @@ dependencies = [
     "PyYAML",
     "psycopg2-binary",
     "entise",
-    "infdb",
+    "pyinfdb",
 ]
 
 [tool.uv.sources]
-infdb = { path = "/workspaces/infdb_package", editable = true }
+pyinfdb = { path = "/workspaces/infdb_package", editable = true }
 entise = { git = "https://github.com/tum-ens/EnTiSe", branch = "develop" }
