+++ tools/infdb-basedata-ways/pyproject.toml
@@ -3,10 +3,10 @@ name = "infdb-basedata-ways"
 version = "1"
 requires-python = "==3.12.*"
 dependencies = [
-    "infdb",
+    "pyinfdb",
     "geopandas",
     # Add your packages here
 ]
 
 [tool.uv.sources]
-infdb = { path = "/workspaces/infdb_package", editable = true }
+pyinfdb = { path = "/workspaces/infdb_package", editable = true }
