### 2026-06-10 890d3ae
Remove unused example files in basedata
