+++ tools/infdb-basedata-ways/src/choose_a_name.py
@@ -1,15 +1,0 @@
-"""
-Template module for .
-
-This module provides an empty file to add your own code and functions
-"""
-
-
-def example_function(variable):
-    """
-    Example function that returns the input variable.
-    """
-
-    # add your code here
-
-    return variable
