### 2026-06-10 f801c09
Simplify pylovo-generation env files
