### 2026-06-10 ad242b8
Update pylovo and its generation yaml
