+++ tools/pylovo-generation/configs/config_generation.yaml
@@ -1,0 +1,99 @@
+# =============================================================================
+# Version CONFIGURATION
+# =============================================================================
+VERSION_ID: "1" # Change version if you changed any grid parameters or queries (e.g. transformer query)
+VERSION_COMMENT: default # State here which parameters, queries are used to compute the grids with current VERSION_ID
+
+# =============================================================================
+# EXECUTION CONFIGURATION
+# =============================================================================
+PARALLEL: True # Toggles whether the grid generation uses multiple workers in parallel
+N_JOBS_PERCENT: 30  # Percentage of CPU cores to use for parallel grid generation
+ANALYZE_GRIDS: True # Toggles analysis after grid generation (basic parameters, trafo analysis, cable counting)
+LOG_LEVEL: INFO #DEBUG, INFO, ERROR - Logging configurations
+SAVE_GRID_FOLDER: False # Toggles whether the grid json files will be saved in a folder or just in the db
+RESULT_DIR: results # Directory for storing analysis results and outputs
+
+# =============================================================================
+# GRID GENERATION CONFIGURATION
+# =============================================================================
+ELECTRICAL_BACKEND: pandapower # Electrical backend configuration - Options: 'pandapower', 'opendss'
+### Electrical load calculation parameters
+PEAK_LOAD_HOUSEHOLD: 17.425 # default: 14.5kW, around 15% of buildings have 30 kW with electric hot water supply
+SIM_FACTOR: {Residential: 0.07, Public: 0.6, Commercial: 0.5} # Simultaneous load factor - formula: simultaneous_load(n) = peak_load * (sim_factor + (1 - sim_factor) * (n ** (-3 / 4)))
+DEFAULT_POWER_FACTOR: 0.95  # cos(phi) for residential loads
+
+### Consumer categories for load calculation
+CONSUMER_CATEGORIES:
+  - {consumer_category_id: 1, definition: Commercial, peak_load: null, yearly_consumption: null, peak_load_per_m2: 79.00, yearly_consumption_per_m2: 245.70, sim_factor: 0.50}
+  - {consumer_category_id: 2, definition: Public, peak_load: null, yearly_consumption: null, peak_load_per_m2: 29.00, yearly_consumption_per_m2: 155.70, sim_factor: 0.60}
+  - {consumer_category_id: 3, definition: SFH, peak_load: PEAK_LOAD_HOUSEHOLD, yearly_consumption: 0.00, peak_load_per_m2: null, yearly_consumption_per_m2: null, sim_factor: 0.07}
+  - {consumer_category_id: 4, definition: MFH, peak_load: PEAK_LOAD_HOUSEHOLD, yearly_consumption: 0.00, peak_load_per_m2: null, yearly_consumption_per_m2: null, sim_factor: 0.07}
+  - {consumer_category_id: 5, definition: TH, peak_load: PEAK_LOAD_HOUSEHOLD, yearly_consumption: 0.00, peak_load_per_m2: null, yearly_consumption_per_m2: null, sim_factor: 0.07}
[…]
