### 2026-06-11 c2b450d
Update basedata docs
