+++ docs/mkdocs/docs/tools/infdb-basedata-buildings/index.md
@@ -1,15 +1,5 @@
 # InfDB-basedata-buildings
-The tool `InfDB-basedata-buildings` processes building-related data as fundamental data basis for various applications and analyses.
-
-<!-- ## Contents
-- Objective (Scope, Motivation)
-- Architecture (Design, Implementation)
-    - Structure (Project/Code)
-    - Data Pipeline
-    - Code (Classes and Functions)
-    - Dependencies
-- Usage (Quick Start, Requirements, Configuration) -->
-
+The `basedata-buildings` tool combines LOD2 building data with census statistics to generate an enriched building dataset. The tool performs data cleaning, attribute filling, population disaggregation, building type classification, and construction period assignment. The processing is carried out through a combination of Python-based orchestration and SQL-based transformation steps within InfDB.
 
 ## Usage
 Details on how to use the tool can be found in the [Usage](usage.md) section.
