### 2026-06-13 17895b9
all timeseries storage functions are moved into entise and ro-heat tool is adjusted respectively
