### 2026-06-18 87a60c0
lizmap version configuration updated
