### 2026-06-18 d00debf
lizmap version configuration updated
