### 2026-06-19 97c3046
udpate config parameters
