### 2026-06-19 78c6a01
adjust config and move stream_chunk_size paramter into ro-heat
