+++ joss/paper.md
@@ -72,19 +72,19 @@ bibliography: paper.bib
 
 It bundles a PostgreSQL [@stonebraker1986design] database extended for geospatial, time-series, and graph data as well as standardized REST [@fielding2000architectural] and OpenAPI standards by the Open Geospatial Consortium (OGC) [@ogc_wfs]. Finally, it offers a configurable import service transforming raw public data into structured, version-controlled schemas.
 
-By decoupling data ingestion, storage, and access from downstream analysis, modeling and planning tools, `InfDB` reduces data pre-processing effort and enables reproducible, transferable energy system modelling workflows across regions and projects. In this way, it can be used for many different applications in energy system modelling, such as district heating planning, electrical distribution network analysis, or urban energy demand estimation.
+By decoupling data ingestion, storage, and access from downstream analysis, modeling and planning tools, `InfDB` reduces data preprocessing effort and enables reproducible, transferable energy system modeling workflows across regions and projects. In this way, it can be used for many different applications in energy system modeling, such as district heating planning, electrical distribution network analysis, or urban energy demand estimation.
   
 # Statement of need
-The transition to a climate-neutral energy system is a central pillar of energy policy, exemplified by Germany's aim for climate neutrality by 2045. New legislative frameworks, such as the requirement for municipal heat planning [@GesetzFurWarmeplanung] and the requirement for grid expansion plans based on regional transition pathway scenarios defined in the German Energy Industry Act [@14dEnWGEinzelnorm], demand that municipalities and Distribution System Operators (DSOs) process vast amounts of energy and infrastructure data.
+The transition to a climate-neutral energy system is a central pillar of energy policy, exemplified by Germany's aim for climate neutrality by 2045. New legislative frameworks, such as the requirement for municipal heat planning [@GesetzFurWarmeplanung] and the requirement for grid expansion plans based on regional transition pathway scenarios defined in the German Energy Industry Act ("Gesetz über die Elektrizitäts- und Gasversorgung") [@14dEnWGEinzelnorm], demand that municipalities and Distribution System Operators (DSOs) process vast amounts of energy and infrastructure data.
 
-However, the current landscape of energy data is fragmented. While the Open Data Strategy of the Federal German Government [@OpenDataStrategieBundesregierung] has increased data availability, this data is published by disparate authorities on different platforms in varying formats, spatial resolutions, and licensing structures. Consequently, energy modeling workflows often suffer from:
+However, the current landscape of energy data is fragmented. While the Open Data Strategy of the German Federal Government [@OpenDataStrategieBundesregierung] has increased data availability, this data is published by disparate authorities on different platforms in varying formats, spatial resolutions, and licensing structures. Consequently, energy modeling workflows often suffer from:
 
-1. **High Pre-processing Effort:** Researchers and professionals spend disproportionate time acquiring raw data before analysis.
+1. **High Preprocessing Effort:** Researchers and professionals spend a disproportionate amount of time acquiring raw data before analysis.
 2. **Limited Workflow Transferability:** Data processing workflows often require substantial adaptation across regions due to differing data formats, interfaces, and conventions.
 3. **Lack of Reproducibility:** Planning results are often one-off studies that are difficult to update or audit.
 4. **Siloed Infrastructure:** DSOs and municipalities lack standardized tools to efficiently manage and share energy and infrastructure data, such as distribution network or building data.
 
-`InfDB` addresses these challenges by providing a reproducible, version-controlled, and automated ETL (Extract, Transform, Load) pipeline. It acts as a middleware between raw public data and high-level energy modelling and planning tools, ensuring that input data is transparent, traceable, and easily updatable. 
[…]
