+++ services/infdb-import/src/utils.py
@@ -974,178 +974,48 @@ def get_clip_geometries_per_scope(target_crs: int, infdb: InfDB):
     return results
 
 
-def create_building_lod2_table(object_id_prefix: str, infdb: InfDB) -> None:
-    """
-    Creates the flat building_lod2 table for the specified object_id_prefix
-    by filtering the source data based on AGS codes.
+def create_building_tables(infdb: InfDB) -> None:
+    """Creates {schema}.building_view and {schema}.building_surface directly from
+    citydb using the fortiss method.
 
-    :param object_id_prefix: Object ID prefix (e.g., "DEBY" for Bavaria, "DENW" for North Rhine-Westphalia, "DEBW" for Baden-Württemberg)
-    :type object_id_prefix: str
-    :param infdb: instance of InfDB for database access and logging
-    :type infdb: InfDB
-    """
-    log = infdb.get_worker_logger()
-
-    # DE means "run all supported region-specific implementations"
-    if object_id_prefix == "DE":
-        create_building_lod2_table("DEBY", infdb)
-        create_building_lod2_table("DENW", infdb)
-        create_building_lod2_table("DEBW", infdb)
-        return
-
-    match object_id_prefix:
-        case "DEBY":
-            ags_id = "09"
-        case "DENW":
-            ags_id = "05"
-        case "DEBW":
-            ags_id = "08"
-        case _:
-            log.error(f"Region {object_id_prefix} not supported for building_lod2.sql")
-            sys.exit(1)
-
-    ags_list = fetch_scope_ags_from_db(infdb)
-    ags_filtered = [s for s in ags_list if s.startswith(ags_id)]
-
-    if ags_filtered:
-
-        def fmt(lst):
-            return ",".join(f"'{s}'" for s in lst)
-
-        table_name = (
-            infdb.get_config_value(
-                [infdb.get_toolname(), "sources", "lod2", "table_name"]
-            )
-            + "_lod2"
-        )
-
-        TEMP_OUTPUT_SCHEMA = "tmp_bld"
-        TEMP_TABLE_NAME = f"{table_name}_{object_id_prefix.lower()}"
-
-        try:
-            with infdb.connect() as db:
-                # # Create central building table
-                # db.execute_sql_file(
-                #     "sql/create_building_table.sql", {"output_schema": output_schema, "table_name": table_name}
-                # )
-                # log.info(f"Created central building_lod2: {output_schema}.{table_name}")
-
-                # Create building table for the region
-                log.info(f"building_lod2: starting {TEMP_OUTPUT_SCHEMA}.{TEMP_TABLE_NAME} ({ags_id}...)")
-                db.execute_sql_file(
-                    "sql/bld.sql",
-                    {
-                        "output_schema": TEMP_OUTPUT_SCHEMA,
-                        "table_name": TEMP_TABLE_NAME,
-                        "ags": fmt(ags_filtered),
-                        "ags_id": ags_id,
-                        "object_id_prefix": object_id_prefix,
-                    },
-                )
-                log.info(f"{TEMP_OUTPUT_SCHEMA}.{TEMP_TABLE_NAME} completed")
-
-        except Exception:
-            infdb.get_logger().exception(f"{TEMP_OUTPUT_SCHEMA}.{TEMP_TABLE_NAME} failed")
-
-
-def create_building_surface_table(infdb: InfDB) -> None:
-    """
-    Creates the flat building_lod2 table for the specified object_id_prefix
-    by filtering the source data based on AGS codes.
-
-    :param object_id_prefix: Object ID prefix (e.g., "DEBY" for Bavaria, "DENW" for North Rhine-Westphalia)
-    :type object_id_prefix: str
[…]
