### 2026-07-09 684e055
Lizmap documentation update, buildings table constraints and logging, qgis version 3.44
