### 2026-07-09 038480f
add published JOSS publication to repository
