+++ CHANGELOG.md
@@ -1,0 +1,20 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+### Added
+- `SECURITY.md` with private vulnerability reporting instructions
+- `CHANGELOG.md` (this file)
+- JOSS paper citation in `README.md`, `CITATION.cff` (`preferred-citation`), and the documentation
+
+## [4.1.4] - 2026-06-09
+
+- Release archived on Zenodo: https://doi.org/10.5281/zenodo.20609117
+- Described in the JOSS paper: Buchenberg et al., (2026). InfDB: An Open Source Energy and Infrastructure Data Ecosystem for Modeling and Planning. *Journal of Open Source Software*, 11(122), 10458. https://doi.org/10.21105/joss.10458
+
+For changes prior to this file's introduction, see the [commit history](https://github.com/tum-ens/InfDB/commits/).
