+++ .github/workflows/tests.yml
@@ -2,7 +2,11 @@ name: tests
 
 on:
   push:
+    branches: [main, develop]
   pull_request:
+
+permissions:
+  contents: read
 
 jobs:
   test:
