### 2026-07-10 6eca348
lizmap documentation update
