+++ services/infdb-import/src/opendata_bavaria.py
@@ -3,20 +3,11 @@ import logging
 import multiprocessing as mp
 import subprocess
 from pathlib import Path
-from typing import Dict, List
 import sys
-import psycopg2
 
-<<<<<<< HEAD
 from pyinfdb import InfDB
 from sqlalchemy import text
 import shlex
-=======
-import geopandas as gpd
-from infdb import InfDB
-from sqlalchemy import text
-from shapely import wkt as shapely_wkt
->>>>>>> f30ea441 (resampling logic is parallelized)
 
 from . import utils
 from .lod2 import _build_urls_for_region
@@ -43,84 +34,13 @@ def _get_gpkg_layers(gpkg: Path) -> list[str]:
         out = subprocess.check_output(["ogrinfo", "-ro", "-q", str(gpkg)], text=True, stderr=subprocess.STDOUT)
         return [line.split(":", 1)[1].split("(")[0].strip() for line in out.splitlines() if ":" in line and "(" in line]
 
-def _write_file_list(file_paths: list[str], list_path: Path) -> None:
+
+
+def _write_file_list(file_paths: list[Path], list_path: Path) -> None:
     """Write one file path per line for GDAL input_file_list usage."""
     with open(list_path, "w", encoding="utf-8") as f:
         for path in file_paths:
             f.write(f"{path}\n")
-
-
-def _cleanup_paths(paths: list[Path], log) -> None:
-    """Delete generated helper/output files if they exist."""
-    for path in paths:
-        try:
-            path.unlink(missing_ok=True)
-        except Exception:
-            log.warning("Could not remove temporary file: %s", path)
-
-
-def _resample_dgm1_tile(
-    infdb: InfDB,
-    source_tif: Path,
-    output_tif: Path,
-    source_srid: int,
-    target_res: float,
-) -> Path | None:
-    """Resample one DGM1 tile to the configured target resolution without clipping."""
-    log = infdb.get_worker_logger()
-
-    if output_tif.exists() and output_tif.stat().st_size > 0:
-        return output_tif
-
-    rc = utils.do_cmd(
-        infdb,
-        [
-            "gdalwarp",
-            "-overwrite",
-            "-of",
-            "GTiff",
-            "-co",
-            "TILED=YES",
-            "-co",
-            "COMPRESS=DEFLATE",
-            "-co",
-            "PREDICTOR=2",
-            "-co",
-            "BIGTIFF=IF_SAFER",
-            "-co",
-            "BLOCKXSIZE=512",
-            "-co",
-            "BLOCKYSIZE=512",
-            "-r",
-            "bilinear",
-            "-multi",
-            "-wo",
-            "NUM_THREADS=ALL_CPUS",
-            "-t_srs",
-            f"EPSG:{source_srid}",
-            "-tr",
-            str(target_res),
-            str(target_res),
-            "-tap",
-            "-srcnodata",
-            "-9999",
-            "-dstnodata",
-            "-9999",
-            str(source_tif),
-            str(output_tif),
-        ],
-    )
-
-    if rc != 0:
-        log.error("DGM1: gdalwarp resampling failed for %s", source_tif)
-        return None
-
-    if not output_tif.exists() or output_tif.stat().st_size == 0:
-        log.warning("DGM1: resampled TIFF is empty for %s", source_tif)
-        return None
-
-    return output_tif
-
 
 def _chunk_list(items: list, batch_size: int) -> list[list]:
     return [
@@ -129,56 +49,65 @@ def _chunk_list(items: list, batch_size: int) -> list[list]:
     ]
 
 
-def _resample_dgm1_tile_worker(
+def _convert_dgm1_tile_to_cog_worker(
     source_tif_str: str,
[…]
