+++ services/infdb-http/docker-compose.yml
@@ -1,0 +1,20 @@
+services:
+  infdb-http:
+    build: .
+    restart: unless-stopped
+    ports:
+      - "8080:80"
+    environment:
+      - BASIC_AUTH_USER=${BASIC_AUTH_USER:-infdb_user}
+      - BASIC_AUTH_PASSWORD=${BASIC_AUTH_PASSWORD:-password}
+    volumes:
+      - ./data:/usr/share/nginx/html
+      # Serve the imported rasters read-only so GDAL clients can open the COGs
+      # over HTTP with /vsicurl/ instead of downloading them.
+      - infdb-import-data:/usr/share/nginx/html/import:ro
+
+volumes:
+  # Same declaration as services/infdb-import/compose.yml, so both projects
+  # attach to the one volume named infdb-import-data.
+  infdb-import-data:
+    name: infdb-import-data
