### 2026-07-24 b79ab1c
Remove infdb-fdw-setup tool
