+++ tools/infdb-fdw-setup/pyproject.toml
@@ -1,12 +1,0 @@
-[project]
-name = "infdb-fdw-setup"
-version = "1"
-requires-python = "==3.12.*"
-dependencies = [
-    "infdb",
-    "geopandas",
-    # Add your packages here
-]
-
-[tool.uv.sources]
-infdb = { path = "/workspaces/infdb_package", editable = true }
