### 2026-07-27 4aac601
overview target resolution logic is updated
