+++ services/infdb-import/src/opendata_bavaria.py
@@ -79,7 +79,9 @@ def _convert_dgm1_tile_to_cog_worker(
                 "-of",
                 "COG",
                 "-co",
-                "COMPRESS=ZSTD",
+                "COMPRESS=DEFLATE",
+                "-co",
+                "PREDICTOR=2",
                 "-co",
                 "BLOCKSIZE=512",
                 "-r",
@@ -100,7 +102,9 @@ def _convert_dgm1_tile_to_cog_worker(
                 "-of",
                 "COG",
                 "-co",
-                "COMPRESS=ZSTD",
+                "COMPRESS=DEFLATE",
+                "-co",
+                "PREDICTOR=2",
                 "-co",
                 "BLOCKSIZE=512",
                 "-a_nodata",
@@ -265,7 +269,9 @@ def _build_consolidated_overview(
             "-co",
             "TILED=YES",
             "-co",
-            "COMPRESS=ZSTD",
+            "COMPRESS=DEFLATE",
+            "-co",
+            "PREDICTOR=2",
             "-co",
             "BIGTIFF=YES",
             "-co",
