### 2026-07-29 30eab24
fix infdb.sh bug for infdb-import
