### 2026-07-29 90367be
osm data is integrated inro infdb-import
