+++ services/infdb-import/src/osm.py
@@ -1,0 +1,310 @@
+import logging
+import subprocess
+import time
+
+from pyinfdb import InfDB
+from sqlalchemy import text
+
+from . import utils
+
+# Module logger (replaced by the worker logger inside load()).
+log = logging.getLogger(__name__)
+
+
+# ====================================================================================
+# OpenStreetMap loader
+# ====================================================================================
+#
+# Behaves like the other loaders: on every import run it checks whether the OSM data
+# is already present, and if not, downloads + imports it first, then clips to scope.
+#
+#   1. DOWNLOAD + FULL IMPORT (only if missing)
+#      Launches the tested pgosm-flex container (osm2pgsql under the hood) as a
+#      SIBLING container via the host Docker daemon. pgosm-flex downloads the WHOLE
+#      subregion (e.g. Bavaria) from Geofabrik and imports it - in external Postgres
+#      mode - directly into a STAGING schema (default "osm") of the infdb database.
+#      This is the exact two-step pgosm-flex workflow (docker run -d, then
+#      docker exec ... pgosm_flex.py), only the DB target is the infdb database.
+#      It is heavy but runs only once: the staging schema persists across runs
+#      (main.py drops only 'opendata' / 'tmp_bld', never the staging schema).
+#
+#   2. SCOPE CLIP (every run)
+#      Copies ONLY the features inside the configured scope
+#      (config-infdb-import.yml -> top-level 'scope:') from the staging schema into
+#      the 'opendata' schema (opendata.osm_<table>).
+#
+# Requirements (see Dockerfile / compose.yml): the docker CLI is installed in this
+# image and the host Docker socket is mounted, so this loader can launch pgosm-flex.
+# ====================================================================================
+
+_PGOSM_IMAGE = "rustprooflabs/pgosm-flex"
+_PGOSM_CONTAINER = "infdb_pgosm"
+_PGOSM_VOLUME = "pgosm-data"  # caches the downloaded .osm.pbf between runs
+
+
+# ------------------------------------------------------------------------------------
+# Staging import (pgosm-flex) helpers
+# ------------------------------------------------------------------------------------
+
+
+def _staging_is_populated(conn, schema: str) -> bool:
+    """True once pgosm-flex has finished: it writes a metadata table <schema>.pgosm_flex."""
+    return bool(
+        conn.execute(
+            text(
+                """
+                SELECT 1 FROM information_schema.tables
+                WHERE table_schema = :s AND table_name = 'pgosm_flex'
+                """
+            ),
+            {"s": schema},
+        ).scalar()
+    )
+
+
+def _wait_for_internal_pg(name: str, log, tries: int = 45, delay: int = 2) -> None:
+    """Wait until the pgosm container's internal Postgres is up (needed before exec)."""
+    for _ in range(tries):
+        rc = subprocess.run(
+            ["docker", "exec", name, "pg_isready", "-q"],
+            stdout=subprocess.DEVNULL,
+            stderr=subprocess.DEVNULL,
+        ).returncode
+        if rc == 0:
+            return
+        time.sleep(delay)
+    log.warning("OSM: pgosm internal Postgres not ready after %ds; proceeding anyway.", tries * delay)
+
+
+def _run_pgosm_import(infdb: InfDB, cfg: dict) -> None:
+    """Run the two-step pgosm-flex workflow against the infdb database (external mode)."""
+    log = infdb.get_worker_logger()
+    p = infdb.get_db_parameters_dict()
+
+    staging_schema = cfg["staging_schema"]
+
+    # Remove any stale container left over from a previous failed run.
+    utils.do_cmd(infdb, ["docker", "rm", "-f", _PGOSM_CONTAINER])
+
[…]
