### 2026-07-30 8d5ede8
fix bug in run uv in pyproject.toml
