+++ pyproject.toml
@@ -10,16 +10,16 @@ authors = [
 ]
 requires-python = "==3.12.*"
 
+dependencies = [
+  "pyinfdb",
+  "geopandas",
+]
+
 [project.urls]
 Homepage = "https://github.com/tum-ens/InfDB"
 Documentation = "https://tum-ens.github.io/InfDB"
 Issues = "https://github.com/tum-ens/InfDB/issues"
 Paper = "https://doi.org/10.21105/joss.10458"
-
-dependencies = [
-    "pyinfdb",
-    "geopandas",
-]
 
 [tool.uv.sources]
 pyinfdb = { path = "src/infdb_package", editable = true }
@@ -61,9 +61,12 @@ test = [
 ]
 
 
-#[build-system]
-#requires = ["hatchling"]
-#build-backend = "hatchling.build"
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/infdb_package"]
 
 [tool.ruff]
 lint.select = ["E", "F", "B", "I"]
