### 2026-07-30 b2f4870
remove docker dependency on infdb-import and add connection check within docker on startup in python
