### 2026-07-30 850656c
make fdw conformant to utils.if_active standard
