### 2026-07-30 a972108
split licence and citationi in Readme.md
