### 2026-08-03 5556129
adjust commenting for osm implementation
