### 2026-08-05 0204886
state-level fallback for AGS with no census construction year data
