+++ tools/ro-heat/configs/config-ro-heat.yml
@@ -28,7 +28,7 @@ ro-heat:
             simulation_year: 2023
             heating_setpoint: 20.0 # heating setpoint in °C
             random_seed: 42 # explain this value
-            method: 1R1C # 1R1C or 1R0C or 1R0C_internal
+            method: 1R0C_internal # 1R1C or 1R0C or 1R0C_internal
             Heated_Area_Ratio: 0.8 # ratio of heated area to total area (0-1)
         refurbishment:
             outer_wall:
