### 2026-08-05 bc972eb
DWD data is imported into need_intern schema
