+++ services/infdb-import/main.py
@@ -7,6 +7,7 @@ from src import (
     basemap,
     bkg,
     census2022,
+    dwd,
     gebaeude_neuburg,
     infdb_fdw_setup,
     kwp_nrw,
@@ -109,23 +110,8 @@ def main() -> None:
     # Launch data loading in parallel
     mp.freeze_support()
     processes: List[mp.Process] = []
-    processes.append(mp.Process(target=_run_loader, args=(need.load,), name="need"))
-    processes.append(mp.Process(target=_run_loader, args=(tabula.load,), name="tabula"))
-    processes.append(mp.Process(target=_run_loader, args=(plz.load,), name="plz"))
-    processes.append(mp.Process(target=_run_loader, args=(basemap.load,), name="basemap"))
-    processes.append(mp.Process(target=_run_loader, args=(census2022.load,), name="census2022"))
-    processes.append(mp.Process(target=_run_loader, args=(openmeteo.load,), name="openmeteo"))
-    processes.append(mp.Process(target=_run_loader, args=(kwp_nrw.load,), name="kwp_nrw"))
-    processes.append(mp.Process(target=_run_loader, args=(kwp_nrw_oberhausen.load,), name="kwp_nrw_oberhausen"))
-    processes.append(mp.Process(target=_run_loader, args=(gebaeude_neuburg.load,), name="gebaeude-neuburg"))
-    processes.append(
-        mp.Process(target=_run_loader, args=(waermeatlas_hessen_bensheim.load,), name="waermeatlas_hessen_bensheim")
-    )
-    # processes.append(mp.Process(target=_run_loader, args=(wetterdienst.load,), name="wetterdienst"))
-    processes.append(mp.Process(target=_run_loader, args=(nrw_opencloud.load,), name="nrw_opencloud"))
-    processes.append(mp.Process(target=_run_loader, args=(opendata_bavaria.load,), name="opendata_bavaria"))
-    processes.append(mp.Process(target=_run_loader, args=(osm.load,), name="osm"))
-    processes.append(mp.Process(target=_run_loader, args=(lod2.load,), name="lod2"))
+    
+    processes.append(mp.Process(target=_run_loader, args=(dwd.load,), name="dwd"))
 
     for process in processes:
         process.start()
