### 2026-08-05 1377b75
scope filtering for DWD doesn't work so it is disabled
