### 2026-08-05 03fb936
get_gpkg_layers is moved into utils.py
