### 2026-08-15 c6b253b
refactor ro_heat and linear-heat-density to run in parallel on ags level
