### 2026-08-17 e7707c9
for our osm use case, routing is not used so the pgosm schema is dropped
