### 2026-08-18 9e9d617
Uncomment dhmin from develop compose.yml
