### 2026-08-19 2a912f5
ro-heat timeseries storage also supports parallel ags processing
