### 2026-08-26 76c8724
mixed use buildings are identifed and residential distribution is adjusted
