### 2026-08-27 c078dae
temporary: filter out 16_logging_and_constraints_for_lizmap.sql
