### 2026-09-09 2e875c2
Hot fix to avoid
$ bash infdb.sh import
=== Importing data ===
no service selected

error
