### 2026-09-21 662aab3
fix: removed external:true docker compose manages now container during setup
